Reworked the api with cursor
