+++ package.json
@@ -28,7 +28,6 @@
     "@tanstack/react-query": "^5.62.7",
     "@vercel/analytics": "^1.5.0",
     "@vercel/node": "^5.6.3",
-    "appwrite": "^16.0.2",
     "axios": "^1.7.9",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
