+++ src/_root/pages/PlaylistPage.tsx
@@ -28,9 +28,9 @@ import LoaderMusic from "@/components/shared/loaderMusic";
 import PlaylistEntry from "@/components/PlaylistEntry";
 import { usePlayerContext } from "@/context/PlayerContext";
 import { useUserContext } from "@/lib/AuthContext";
-import { getPlaylistById, processPlaylistCover, updatePlaylistMetadata, addItemToPlaylist, updatePlaylistCover, addSongToDatabase, updateItemsOrder, deletePlaylist, removeItemFromPlaylist, addAlbumComplex, batchSyncSongsToDatabase, batchAddItemsToPlaylist, getAlbumTracks} from "@/lib/appwrite/api";
+import { getPlaylistById, processPlaylistCover, updatePlaylistMetadata, addItemToPlaylist, updatePlaylistCover, addSongToDatabase, updateItemsOrder, deletePlaylist, removeItemFromPlaylist, addAlbumComplex, batchSyncSongsToDatabase, batchAddItemsToPlaylist, getAlbumTracks} from "@/lib/supabase/api";
 import { Search } from "lucide-react";
-import { searchSpotify, getSpotifyToken, SpotifyTrackById, getSpotifyPlaylistTracks } from "@/lib/appwrite/spotify";
+import { searchSpotify, getSpotifyToken, SpotifyTrackById, getSpotifyPlaylistTracks } from "@/lib/integrations/spotify";
 import { useToast } from "@/hooks/use-toast";
 import { DragDropContext, Droppable, Draggable, DropResult } from "@hello-pangea/dnd";
 import { Track } from "@/types";
