+++ src/components/ReviewItem.tsx
@@ -7,13 +7,12 @@ import {
   addLikeToReview,
   removeLikeFromReview,
   checkIfUserLikedReview,
-} from "@/lib/appwrite/api";
+} from "@/lib/supabase/api";
 import { Heart } from "lucide-react";
 import AuthModal from "@/components/shared/AuthModal"; // Import your modal
 
 const ReviewItem = (review: SongReview | AlbumReview) => {
 
-  console.log("Rendering ReviewItem with review:", review); // Debug log to check review data
   const [showFull, setShowFull] = useState(false);
   const [liked, setLiked] = useState(false);
   const [likeCount, setLikeCount] = useState(review.likes ?? 0);
