+++ src/components/shared/LogEntryModal.tsx
@@ -1,8 +1,8 @@
 import { useState, useEffect, useRef } from "react";
 import { motion } from "framer-motion";
 import { Edit2, X, Search, Loader2 } from "lucide-react";
-import { getSpotifyToken, searchSpotify, SpotifyTrackById } from "@/lib/appwrite/spotify";
-import { addUpdateRatingAlbum, addUpdateRatingSong, addAlbumComplex, addSongToDatabase, addReviewAlbum, addReviewSong } from "@/lib/appwrite/api";
+import { getSpotifyToken, searchSpotify, SpotifyTrackById } from "@/lib/integrations/spotify";
+import { addUpdateRatingAlbum, addUpdateRatingSong, addAlbumComplex, addSongToDatabase, addReviewAlbum, addReviewSong } from "@/lib/supabase/api";
 import { useUserContext } from "@/lib/AuthContext";
 import StarIcon from "@/components/shared/StarIcon";
 import { Input } from "@/components/ui/input";
