--- src/lib/appwrite/api.ts
@@ -1,3337 +1,0 @@
-import { nanoid } from "nanoid";
-
-import { SuggestedAlbum, AlbumDetails, ArtistDetails, INewUser, IUser, IFollow, Listened, RatingGeneral, Comment, Review, Song, SongDetails, SpotifyAlbum, SpotifyAlbumWithTracks, SpotifySong, Activity, ISearchUser, SpotifyTrack, SpotifyArtistDetailed, TrendingResponse, Playlist, SongActivity, AlbumActivity, ReviewActivity } from "@/types";
-import { supabase } from "@/lib/supabaseClient";
-import { getDeezerPreview } from "./deezer";
-
-function normalizeReleaseDate(dateStr: string): string | null {
-    if (!dateStr) return null;
-
-    // year only
-    if (/^\d{4}$/.test(dateStr)) return `${dateStr}-01-01`;
-    // year-month
-    if (/^\d{4}-\d{2}$/.test(dateStr)) return `${dateStr}-01`;
-    // already full
-    if (/^\d{4}-\d{2}-\d{2}$/.test(dateStr)) return dateStr;
-
-    return null; // fallback
-}
-
-export const getProfileUrl = (userId: string): string => {
-    if (!userId) return "/assets/icons/profile-placeholder.svg";
-
-    // This is your project reference from your URL
-    const PROJECT_ID = "akneztqjuwaharlzqwvc"; 
-    
-    // Construct the direct link to the public object
-    return `https://${PROJECT_ID}.supabase.co/storage/v1/object/public/profiles/${userId}/profile.jpg` || "/assets/icons/profile-placeholder.svg";
-};
-
-
-export async function getCurrentUser() {
-    const {
-        data: { session },
-        error,
-    } = await supabase.auth.getSession();
-    if (error) throw error;
-    return session?.user || null;
-}
-
-export const timeAgo = (dateString: string) => {
-    if (!dateString) return "";
-
-    // Ensure UTC interpretation
-    const utcDate = new Date(dateString.endsWith("Z") ? dateString : `${dateString}Z`);
-    const then = utcDate.getTime();
-    const now = Date.now();
-
-    if (isNaN(then)) return "";
-
-    const diff = (now - then) / 1000;
-
-    if (diff < 60) return "just now";
-    if (diff < 3600) return `${Math.floor(diff / 60)}m ago`;
-    if (diff < 86400) return `${Math.floor(diff / 3600)}h ago`;
-    if (diff < 604800) return `${Math.floor(diff / 86400)}d ago`;
-
-    // For older posts, show local formatted date
-    return utcDate.toLocaleString(undefined, {
-        month: "short",
-        day: "numeric"
-    });
-};
-
-
-
-export async function createUserAccount(user: INewUser) {
-    try {
-        // Create auth user in Supabase
-        const { data: authData, error: authError } = await supabase.auth.signUp({
-            email: user.email,
-            password: user.password,
-            options: {
-                data: {
-                    name: user.name,
-                    username: user.username,
-                },
-            },
-        });
-
-        if (authError) throw authError;
-        if (!authData.user) throw new Error("Failed to create user");
-
-        // Save to users table
-        const { data: newUser, error: dbError } = await supabase
-            .from("users")
-            .insert({
-                user_id: authData.user.id,
-                name: user.name,
-                email: user.email,
-                username: user.username,
-            })
-            .select()
-            .single();
-
-        if (dbError) throw dbError;
-
-        return newUser;
-    } catch (error) {
-        console.error("createUserAccount error:", error);
-        return null;
-    }
-}
-
-export async function saveUserToDB(user: {
-    accountId: string;
-    email: string;
-    name: string;
-    username?: string;
-}) {
-    try {
-
-        console.log("adding user to data:", user);
-        const { data: newUser } = await supabase
-            .from("users")
-            .insert(user)
-
-        return newUser
-    } catch (error) {
-        console.log(error);
-    }
-}
-
-function processProfileImage(file: File, maxSize = 500): Promise<Blob> {
-    return new Promise((resolve) => {
-        const img = new Image();
-        img.src = URL.createObjectURL(file);
-        img.onload = () => {
-            // Determine new dimensions
-            let width = img.width;
-            let height = img.height;
-
-            if (width > height) {
-                if (width > maxSize) {
-                    height = (height * maxSize) / width;
-                    width = maxSize;
-                }
-            } else {
-                if (height > maxSize) {
-                    width = (width * maxSize) / height;
-                    height = maxSize;
-                }
-            }
-
-            // Draw to canvas
-            const canvas = document.createElement("canvas");
-            canvas.width = width;
-            canvas.height = height;
-            const ctx = canvas.getContext("2d")!;
-            ctx.drawImage(img, 0, 0, width, height);
-
-            // Convert to JPEG
-            canvas.toBlob((blob) => {
-                resolve(blob!);
-            }, "image/jpeg", 0.9); // 0.9 quality to save some space
-        };
-    });
-}
-
-export async function updateUser({
-    accountId,
-    bio,
-    imageFile,
-    topFive, 
-}: {
-    accountId: string;
-    bio: string;
-    imageFile?: File | null;
-    topFive?: any[]; 
-}) {
-    try {
-        if (imageFile) {
-            const processedImage = await processProfileImage(imageFile);
-            const fileName = `profile.jpg`;
-            const filePath = `${accountId}/${fileName}`;
-
-            const { error: uploadError } = await supabase.storage
-                .from("profiles")
-                .upload(filePath, processedImage, { upsert: true });
-
-            if (uploadError) throw uploadError;
-        }
-
-        // Update user info including the new top_five column
-        const { data, error } = await supabase
-            .from("users")
-            .update({
-                bio,
-                top_five: topFive, 
-            })
-            .eq("user_id", accountId)
-            .select()
-            .single();
-
-        if (error) throw error;
-        return data;
-    } catch (err) {
-        console.error("Error updating user:", err);
-        throw err;
-    }
-}
-
-export async function updateUsername(userId: string, newUsername: string) {
-    try {
-        const { data, error: tableError } = await supabase
-        .from("users") 
-        .update({ username: newUsername })
-        .eq("user_id", userId)
-        .select()
-        .single();
-
-    if (tableError) throw tableError;
-
-    const { error: authError } = await supabase.auth.updateUser({
-        data: { username: newUsername }
-    });
-
-    if (authError) console.error("Auth metadata sync failed, but table updated.");
-
-    return data;
-    } catch (error) {
-        throw error;
-    }
-}
-
-export async function searchUsers(query: string): Promise<ISearchUser[]> {
-    try {
-        if (!query) return [];
-
-        // Search in username or name columns
-        const { data: usersData, error } = await supabase
-            .from("users")
-            .select("user_id, username, name")
-            .or(`username.ilike.%${query}%,name.ilike.%${query}%`)
-            .limit(20);
-
-        if (error) throw error;
-        if (!usersData) return [];
-
-        const results = usersData.map((user: any) => ({
-            id: user.user_id,
-            username: user.username,
-            name: user.name,
-            avatar_url: getProfileUrl(user.user_id),
-        }))
-        
-
-        console.log(results);
-
-        return results;
-    } catch (error) {
-        console.error("Failed to search users:", error);
-        return [];
-    }
-}
-
-
-
-export async function addFollow(followingId: string, followerId: string): Promise<IFollow | null> {
-    const now = new Date();
-    const isoString = now.toISOString();
-    try {
-        const { data, error } = await supabase
-            .from("followers")
-            .insert({
-                following_id: followingId,
-                follower_id: followerId,
-                follow_date: isoString,
-            });
-        if (error) throw error;
-        return data;
-    }
-    catch (error) {
-        console.log(error);
-        return null;
-    }
-}
-
-export async function removeFollow(followingId: string, followerId: string): Promise<boolean> {
-    try {
-        const { error } = await supabase
-            .from("followers")
-            .delete()
-            .eq("following_id", followingId)
-            .eq("follower_id", followerId);
-        if (error) throw error;
-        return true;
-    } catch (error) {
-        console.log(error);
-        return false;
-    }
-}
-
-export async function checkIfFollowing(followingId: string, followerId: string): Promise<boolean> {
-    try {
-        const { data, error } = await supabase
-
-            .from("followers")
-            .select("*")
-            .eq("following_id", followingId)
-            .eq("follower_id", followerId);
-        if (error) throw error;
-        return data.length > 0;
-    } catch (error) {
-        console.log(error);
-        return false;
-    }
-}
-
-
-export async function getSongById(songId: string): Promise<Song | null> {
-    try {
-        const { data: songData, error: songError } = await supabase
-            .from("songs")
-            .select("*")
-            .eq("song_id", songId)
-            .single();
-
-
-
-        if (songError) throw songError;
-        if (!songData) throw new Error("Song not found");
-
-
-        const { data: albumData, error: albumError } = await supabase
-            .from("albums")
-            .select("*")
-            .eq("album_id", songData.album_id)
-            .single();
-
-        if (albumError) throw albumError;
-        if (!albumData) throw new Error("Album not found");
-
-        // Validate or map the returned songData to a Song type
-        const song: Song = {
-            songId: songData.song_id,
-            title: songData.title,
-            album: albumData.titel,
-            release_date: albumData.release_date,
-            spotify_url: songData.spotify_url,
-            album_cover_url: albumData.album_cover_url,
-            popularity: songData.popularity
-        };
-
-        return song;
-    } catch (error) {
-        console.error('Failed to fetch song:', error);
-        return null;
-    }
-}
-
-export async function fetchSongs(page = 1, limit = 20): Promise<Song[]> {
-    try {
-        const from = (page - 1) * limit;
-        const to = from + limit - 1;
-
-        const { data: songData, error: songError } = await supabase
-            .from("songs")
-            .select("*, albums:songs_album_id_fkey(album_cover_url)")
-            .range(from, to);
-
-
-
-        if (songError) throw songError;
-        if (!songData) throw new Error("Song not found");
-
-        console.log(songData);
-
-
-
-        return songData.map((song: any) => ({
-            ...song,
-            songId: song.song_id,
-            album_cover_url: song.albums?.album_cover_url ?? null,
-        }));
-    }
-    catch (error) {
-        console.log(error);
-        return []
-    }
-}
-
-
-export async function addReviewSong(songId: string, userId: string, reviewText: string, reviewTitle?: string) {
-    const id = nanoid();
-    const now = new Date();
-    const isoString = now.toISOString();
-    try {
-
-        const { data: review, error: addReview } = await supabase
-            .from("reviews")
-            .insert({
-                review_id: id,
-                song_id: songId,
-                user_id: userId,
-                review_type: "song",
-                review_text: reviewText,
-                review_title: reviewTitle,
-                created_at: isoString,
-            })
-            .select()
-            .single();
-
-        const { data: getListen, error: getListenError } = await supabase
-            .from("song_listens")
-            .select("*")
-            .eq("song_id", songId)
-            .eq("user_id", userId)
-
-        if (getListenError) throw getListenError;
-
-        if (getListen.length === 0) {
-            const { error: listenError } = await supabase
-                .from("song_listens")
-                .insert({ song_id: songId, user_id: userId, listen_date: isoString })
-            if (listenError) throw listenError;
-        }
-        if (addReview) throw addReview;
-
-        if (!review) throw new Error("Failed to add review");
-
-        return review;
-    } catch (error) {
-        console.log(error);
-        return null;
-    }
-}
-
-
-export async function addReviewAlbum(albumId: string, userId: string, reviewText: string, reviewTitle?: string) {
-    const id = nanoid();
-    const now = new Date();
-    const isoString = now.toISOString();
-    try {
-
-        const { data: review, error: addReview } = await supabase
-            .from("reviews")
-            .insert({
-                review_id: id,
-                album_id: albumId,
-                user_id: userId,
-                review_type: "album",
-                review_text: reviewText,
-                review_title: reviewTitle,
-                created_at: isoString,
-            })
-            .select()
-            .single();
-
-        const { data: getListen, error: getListenError } = await supabase
-            .from("album_listens")
-            .select("*")
-            .eq("album_id", albumId)
-            .eq("user_id", userId)
-
-        if (getListenError) throw getListenError;
-
-        if (getListen.length === 0) {
-            const { error: listenError } = await supabase
-                .from("album_listens")
-                .insert({ album_id: albumId, user_id: userId, listen_date: isoString })
-            if (listenError) throw listenError;
-        }
-
-        if (addReview) throw addReview;
-
-        if (!review) throw new Error("Failed to add review");
-
-        return review;
-    } catch (error) {
-        console.log(error);
-        return null;
-    }
-}
-
-export async function getReviewById(reviewId: string): Promise<Review | null> {
-    try {
-        const { data: reviewData, error: reviewError } = await supabase
-            .from("reviews")
-            .select(`
-                *,
-                likes:reviewlikes(*),
-                creator:users(*),
-                comments:reviewcomment(
-                comment_id,
-                parent_id,
-                content,
-                created_at,
-                updated_at,
-                user:users(*)
-                )
-            `)
-            .eq("review_id", reviewId)
-            .single();
-        if (reviewError) throw reviewError;
-        if (!reviewData) throw new Error("Review not found");
-        let coverUrl = "";
-        let name = "";
-        const reviewType: "song" | "album" = reviewData.song_id ? "song" : "album";
-        if (reviewData.album_id) {
-            const { data: album } = await supabase
-                .from("albums")
-                .select("*")
-                .eq("album_id", reviewData.album_id)
-                .single();
-            coverUrl = album?.album_cover_url ?? "";
-            name = album?.title ?? "";
-        } else if (reviewData.song_id) {
-            const { data: song } = await supabase
-                .from("songs")
-                .select("*, album:albums(album_cover_url)")
-                .eq("song_id", reviewData.song_id)
-                .single();
-            coverUrl = song?.album.album_cover_url ?? "";
-            name = song?.title ?? "";
-        }
-
-        const comments: Comment[] = (reviewData.comments ?? []).map((c: any) => ({
-                commentId: c.comment_id,
-                parentId: c.parent_id,
-                text: c.content,
-                createdAt: c.created_at,
-                updatedAt: c.updated_at,
-                creator: {
-                    accountId: c.user.user_id,
-                    name: c.user.name,
-                    username: c.user.username,
-                    email: c.user.email,
-                    imageUrl: getProfileUrl(c.user.user_id),
-                    bio: c.user.bio ?? "",
-                }
-            }));
-        
-
-        comments.sort(
-            (a, b) =>
-                new Date(b.createdAt).getTime() -
-                new Date(a.createdAt).getTime()
-        )
-
-        console.log(comments)
-
-
-        return {
-            reviewId: reviewData.review_id,
-            text: reviewData.review_text,
-            title: reviewData.review_title,
-            id: reviewData.song_id ?? reviewData.album_id,
-            type: reviewType,
-            createdAt: reviewData.created_at,
-            name: name,
-            album_cover_url: coverUrl,
-            likes: reviewData.likes.length,
-            comments: comments,
-            creator: {
-                accountId: reviewData.creator.user_id,
-                name: reviewData.creator.name,
-                username: reviewData.creator.username,
-                email: reviewData.creator.email,
-                imageUrl: await getProfileUrl(reviewData.creator.user_id),
-                bio: reviewData.creator.bio ?? "",
-            }
-        };
-    } catch (error) {
-        console.error('Failed to fetch review:', error);
-        return null;
-    }
-}
-
-
-
-export async function addLikeToReview(reviewId: string, userId: string): Promise<boolean> {
-    const now = new Date();
-    const isoString = now.toISOString();
-    try {
-        const { error } = await supabase
-            .from("reviewlikes")
-            .insert({
-                review_id: reviewId,
-                user_id: userId,
-                like_date: isoString
-            });
-        if (error) throw error;
-        return true;
-    } catch (error) {
-        console.log(error);
-        return false;
-    }
-}
-
-export async function removeLikeFromReview(reviewId: string, userId: string): Promise<boolean> {
-    try {
-        const { error } = await supabase
-            .from("reviewlikes")
-            .delete()
-            .eq("review_id", reviewId)
-            .eq("user_id", userId);
-        if (error) throw error;
-        return true;
-    } catch (error) {
-        console.log(error);
-        return false;
-    }
-}
-
-export async function checkIfUserLikedReview(reviewId: string, userId: string): Promise<boolean> {
-    try {
-        const { data, error } = await supabase
-            .from("reviewlikes")
-            .select("*")
-            .eq("review_id", reviewId)
-            .eq("user_id", userId);
-        if (error) throw error;
-        return data.length > 0;
-    } catch (error) {
-        console.log(error);
-        return false;
-    }
-}
-
-export async function addCommentToReview(reviewId: string, userId: string, text: string, parentId?: string): Promise<boolean> {
-    const id = nanoid();
-    const now = new Date();
-    const isoString = now.toISOString();
-    try {
-        const { error } = await supabase
-            .from("reviewcomment")
-            .insert({
-                review_id: reviewId,
-                user_id: userId,
-                comment_id: id,
-                parent_id: parentId,
-                content: text,
-                created_at: isoString
-            })
-        if (error) throw error;
-        return true;
-    }
-    catch (error) {
-        console.log(error);
-        return false;
-    }
-}
-
-
-
-export async function getUserById(userId: string): Promise<IUser | null> {
-    try {
-
-        const { data: userData, error: userError } = await supabase
-            .from("users")
-            .select("*")
-            .eq("user_id", userId)
-            .single();
-
-        if (userError) throw userError;
-        if (!userData) throw new Error("User not found")
-
-        const { data: followersData, error: followersError } = await supabase
-            .from("followers")
-            .select("follower_id", { count: 'exact' })
-            .eq("following_id", userId);
-
-        const { data: followingData, error: followingError } = await supabase
-            .from("followers")
-            .select("following_id", { count: 'exact' })
-            .eq("follower_id", userId);
-        if (followersError) console.error("Error fetching followers count:", followersError);
-        if (followingError) console.error("Error fetching following count:", followingError);
-
-        const followersCount = followersData ? followersData.length : 0;
-        const followingCount = followingData ? followingData.length : 0;
-
-
-        const user: IUser = {
-            accountId: userData.user_id,
-            name: userData.name,
-            username: userData.username,
-            email: userData.email,
-            imageUrl: await getProfileUrl(userId),
-            bio: userData.bio,
-            followersCount,
-            followingCount, 
-            top_five: userData.top_five
-        };
-
-        return user;
-    } catch (error) {
-        console.error('Failed to fetch user:', error);
-        return null;
-    }
-}
-
-
-export async function getFollowersList(userId: string, type: 'followers' | 'following') {
-    const columnToMatch = type === 'followers' ? 'following_id' : 'follower_id';
-    const columnToSelect = type === 'followers' ? 'follower_id' : 'following_id';
-
-    const { data, error } = await supabase
-        .from("followers")
-        .select(`
-            users!${columnToSelect} (
-                user_id,
-                name,
-                username,
-                bio
-            )
-        `)
-        .eq(columnToMatch, userId);
-
-    if (error) {
-        console.error("Supabase Error:", error);
-        throw error;
-    }
-
-    const usersWithImages = await Promise.all(
-        data.map(async (item: any) => {
-            const user = item.users;
-            if (!user) return null;
-
-            return {
-                ...user,
-                accountId: user.user_id,
-                imageUrl: await getProfileUrl(user.user_id)
-            };
-        })
-    );
-    
-    return usersWithImages;
-}
-
-export async function getNotifications(userId: string) {
-    const { data, error } = await supabase
-        .from('notifications')
-        .select(`
-            id,
-            type,
-            created_at,
-            is_read,
-            entity_id,
-            actor_id,
-            actor:users!notifications_actor_id_fkey ( username )
-        `)
-        .eq('recipient_id', userId)
-        .order('created_at', { ascending: false });
-
-    if (error) throw error;
-
-    return (data || []).map(n => ({
-        id: n.id,
-        type: n.type,
-        createdAt: n.created_at,
-        isRead: n.is_read,
-        entityId: n.entity_id,
-        actorId: n.actor_id,
-        actorUsername: (n.actor as any)?.username || "Someone",
-        actorAvatar: getProfileUrl(n.actor_id)
-    }));
-}
-
-export async function getSongDetailsById(songId: string): Promise<SongDetails | null> {
-    try {
-        // Fetch the song
-        const { data: songData, error: songError } = await supabase
-            .from("songs")
-            .select("*")
-            .eq("song_id", songId)
-            .single();
-
-        if (songError) throw songError;
-        if (!songData) throw new Error("Song not found");
-
-        // Fetch the album
-        const { data: albumData, error: albumError } = await supabase
-            .from("albums")
-            .select("*")
-            .eq("album_id", songData.album_id)
-            .single();
-
-        if (albumError) throw albumError;
-        if (!albumData) throw new Error("Album not found");
-
-        // Fetch artists
-        const { data: artists, error: artistError } = await supabase
-            .from("artistsongs")
-            .select(`artist:artists(*)`)
-            .eq("song_id", songData.song_id);
-
-        if (artistError) throw artistError;
-        if (!artists || artists.length === 0) throw new Error("Artist(s) not found");
-
-        const artistList = artists.map((a) => a.artist);
-
-        // Fetch reviews
-        const { data: reviews, error: reviewError } = await supabase
-            .from("reviews")
-            .select("*, creator:users(*), likes:reviewlikes(*)")
-            .eq("song_id", songId)
-            .order("created_at", { ascending: false });
-
-        if (reviewError) throw reviewError;
-
-        const preview_url = await getDeezerPreview(songData.title, songData.artists, songData.isrc)
-
-        // Build song object
-        const song: SongDetails = {
-            songId: songData.song_id,
-            title: songData.title,
-            album: albumData.title,
-            album_id: albumData.album_id,
-            release_date: albumData.release_date,
-            spotify_url: songData.spotify_url,
-            album_cover_url: albumData.album_cover_url,
-            popularity: songData.popularity,
-            artists: artistList,
-            ratings: [],
-            reviews: await Promise.all(
-                reviews.map(async (r: any) => ({
-                    reviewId: r.review_id,
-                    text: r.review_text,
-                    title: r.review_title,
-                    creator: {
-                        accountId: r.creator.user_id,
-                        name: r.creator.name,
-                        username: r.creator.username,
-                        email: r.creator.email,
-                        imageUrl: await getProfileUrl(r.creator.user_id),
-                        bio: r.creator.bio ?? "",
-
-                    },
-                    song: songData,
-                    likes: r.likes.length,
-                    createdAt: r.created_at,
-                    updatedAt: r.created_at,
-                }))
-            ),
-            isrc: songData.isrc,
-            preview_url: preview_url
-        };
-
-        return song;
-    } catch (error) {
-        console.error("Failed to fetch song:", error);
-        return null;
-    }
-}
-
-
-export async function getArtistDetailsById(artistId: string): Promise<ArtistDetails | null> {
-    try {
-        const { data: artistData, error: artistError } = await supabase
-            .from("artists")
-            .select(`
-                *,
-                artistalbum (
-                albums (*)
-                )
-            `)
-            .eq("artist_id", artistId)
-            .single();
-
-        if (artistError) throw artistError;
-        if (!artistData) throw new Error("Artist not found");
-        if (artistData.fully_loaded == false) return null;
-
-        const albums = (artistData.artistalbum ?? []).map((item: any) => ({
-            albumId: item.albums.album_id,
-            title: item.albums.title,
-            spotify_url: item.albums.spotify_url,
-            album_cover_url: item.albums.album_cover_url,
-            release_date: item.albums.release_date,
-            album_type: item.albums.album_type ?? "",
-        }));
-
-        console.log(albums);
-
-        return {
-            artistId: artistData.artist_id,
-            name: artistData.name,
-            spotify_url: artistData.spotify_url,
-            image_url: artistData.image_url,
-            followers: artistData.followers,
-            genres: [],
-            albums
-        };
-
-    } catch (error) {
-        console.error("Failed to fetch Artist:", error);
-        return null;
-    }
-}
-
-
-export async function getAlbumDetailsById(albumId: string): Promise<AlbumDetails | null> {
-    try {
-        // Fetch the album
-        const { data: albumData, error: albumError } = await supabase
-            .from("albums")
-            .select("*")
-            .eq("album_id", albumId)
-            .single();
-
-        if (albumError) throw albumError;
-        if (!albumData) throw new Error("Album not found");
-
-        if (albumData.fully_loaded === false) return null;
-
-
-
-        // fetch all song information
-        const { data: songData, error: songError } = await supabase
-            .from("songs")
-            .select("*")
-            .eq("album_id", albumId)
-            .order("track_number", { ascending: true });
-
-
-        if (songError) throw songError;
-        if (!songData) throw new Error("No songs found");
-
-        const trackInputs = songData.map(s => ({
-            songId: s.song_id,
-            title: s.title,
-            artist: s.artist, // Assuming s.artist is the primary artist string
-            isrc: s.isrc
-        }));
-
-        const { data: enrichmentData, error: enrichmentError } = await supabase.functions.invoke('enrich-album', {
-            body: { tracks: trackInputs }
-        });
-
-
-        // fetch artists 
-        const { data: artists, error: artistError } = await supabase
-            .from("artistalbum")
-            .select(`
-                artist:artists(*)  
-            `)
-            .eq("album_id", albumId);
-
-        if (artistError) throw artistError;
-        if (!artists || artists.length === 0) throw new Error("Artist(s) not found");
-
-        // Extract the artist info
-        const artistList = artists.map(a => a.artist);
-
-
-
-        // fetch album reviews
-        const { data: reviews, error: reviewError } = await supabase
-            .from("reviews")
-            .select("*, creator:users(*), likes:reviewlikes(*)")
-            .eq("album_id", albumId)
-            .order("created_at", { ascending: false });
-
-        if (reviewError) throw reviewError;
-
-        //get preview url's 
-        const previewMap = new Map();
-        if (!enrichmentError && enrichmentData?.tracks) {
-            enrichmentData.tracks.forEach((t: any) => {
-                previewMap.set(t.songId, t.preview_url);
-            });
-        }
-
-        const album: AlbumDetails = {
-            albumId: albumData.album_id,
-            title: albumData.title,
-            spotify_url: albumData.spotify_url,
-            album_cover_url: albumData.album_cover_url,
-            release_date: albumData.release_date,
-            tracks: songData.map((s: any) => ({
-                songId: s.song_id,
-                title: s.title,
-                album: albumData.title,
-                album_id: albumData.album_id,
-                reviews: [],
-                ratings: [],
-                artists: [],
-                spotify_url: s.spotify_url,
-                album_cover_url: albumData.album_cover_url,
-                release_date: albumData.release_date,
-                popularity: s.pop,
-                isrc: s.isrc, 
-                preview_url: previewMap.get(s.song_id) || null,
-            })),
-            album_type: albumData.album_type,
-            artists: artistList,
-            reviews: await Promise.all(
-                reviews.map(async (r: any) => {
-                    let imageUrl = "";
-                    try {
-                        const { data: signedData, error: signedError } = await supabase.storage
-                            .from("profiles")
-                            .createSignedUrl(`${r.creator.user_id}/profile.jpg`, 60 * 60); // 1 hour
-
-                        if (!signedError && signedData?.signedUrl) {
-                            imageUrl = signedData.signedUrl;
-                        }
-                    } catch (err) {
-                        console.error("Failed to generate signed URL:", err);
-                    }
-
-                    return {
-                        reviewId: r.review_id,
-                        text: r.review_text,
-                        title: r.review_title,
-                        creator: {
-                            accountId: r.creator.user_id,
-                            name: r.creator.name,
-                            username: r.creator.username,
-                            email: r.creator.email,
-                            imageUrl,
-                            bio: r.creator.bio ?? ""
-                        },
-                        album: albumData,
-                        likes: r.likes.length,
-                        createdAt: r.created_at,
-                        updatedAt: r.created_at,
-                    };
-                })
-            ),
-        };
-        return album
-    } catch (error) {
-        console.error("Failed to fetch Album:", error);
-        return null;
-    }
-}
-
-
-export async function getAlbumTrackRatings(
-    albumId: string,
-    userId: string
-): Promise<{ songId: string; rating: number }[]> {
-    const { data, error } = await supabase
-        .from("song_rating")
-        .select("song_id, rating, songs!inner(album_id)")
-        .eq("songs.album_id", albumId)
-        .eq("user_id", userId);
-
-    if (error) {
-        console.error("Error fetching user track ratings:", error);
-        return [];
-    }
-
-    if (!data) return [];
-
-    // Simplify the structure to just song_id and rating
-    return data.map((t) => ({
-        songId: t.song_id,
-        rating: t.rating,
-    }));
-}
-
-
-
-// add song, album artist into the relevant tables depending if they are there or not 
-export async function addSongToDatabase(song: SpotifySong) {
-    try {
-        // 1. Upsert Album (onConflict uses the primary key)
-        const { error: albumError } = await supabase
-            .from("albums")
-            .upsert({
-                album_id: song.album.id,
-                title: song.album.name,
-                spotify_url: song.album.external_urls.spotify,
-                album_cover_url: song.album_cover_url,
-                release_date: normalizeReleaseDate(song.album.release_date),
-                total_tracks: song.album.total_tracks,
-            }, { onConflict: 'album_id' });
-
-        if (albumError) throw albumError;
-
-        // 2. Upsert Song
-        const { error: songInsertError } = await supabase
-            .from("songs")
-            .upsert({
-                song_id: song.songId,
-                title: song.title,
-                album_id: song.album.id,
-                spotify_url: song.spotify_url,
-                pop: song.popularity,
-                isrc: song.isrc,
-            }, { onConflict: 'song_id' });
-
-        if (songInsertError) throw songInsertError;
-
-        // 3. Upsert Artists and link them
-        for (const artist of song.artists) {
-            await supabase.from("artists").upsert({
-                artist_id: artist.id,
-                name: artist.name,
-                spotify_url: artist.external_urls.spotify,
-            }, { onConflict: 'artist_id' });
-
-            await supabase.from("artistsongs").upsert({
-                song_id: song.songId, 
-                artist_id: artist.id 
-            }, { onConflict: 'song_id, artist_id' });
-        }
-
-        return true;
-    } catch (error) {
-        console.error("Error syncing song to database:", error);
-        throw error;
-    }
-}
-
-export async function batchSyncSongsToDatabase(songs: any[]) {
-    try {
-        // 1. Prepare unique Albums
-        const uniqueAlbums = Array.from(new Map(songs.map(s => [s.album.id, {
-        album_id: s.album.id,
-        title: s.album.name,
-        spotify_url: s.album.external_urls.spotify,
-        album_cover_url: s.album.images?.[0]?.url || s.album_cover_url,
-        release_date: normalizeReleaseDate(s.album.release_date),
-        total_tracks: s.album.total_tracks,
-        }])).values());
-
-        // 2. Prepare Songs
-        const songsToUpsert = songs.map(s => ({
-        song_id: s.id,
-        title: s.name || s.title,
-        album_id: s.album.id,
-        spotify_url: s.external_urls?.spotify,
-        pop: s.popularity,
-        isrc: s.external_ids?.isrc,
-        preview_url: s.preview_url,
-        track_number: s.track_number
-        }));
-
-        // 3. Prepare Artists (Flattened unique list)
-        const artistMap = new Map();
-        const artistLinks: any[] = [];
-        
-        songs.forEach(s => {
-        s.artists.forEach((a: any) => {
-            artistMap.set(a.id, {
-            artist_id: a.id,
-            name: a.name,
-            spotify_url: a.external_urls?.spotify
-            });
-            artistLinks.push({ song_id: s.id, artist_id: a.id });
-        });
-        });
-
-        // EXECUTE BATCHES
-        const { error: albErr } = await supabase.from("albums").upsert(uniqueAlbums, { onConflict: 'album_id' });
-        if (albErr) throw albErr;
-
-        const { error: songErr } = await supabase.from("songs").upsert(songsToUpsert, { onConflict: 'song_id' });
-        if (songErr) throw songErr;
-
-        const { error: artErr } = await supabase.from("artists").upsert(Array.from(artistMap.values()), { onConflict: 'artist_id' });
-        if (artErr) throw artErr;
-
-        const { error: linkErr } = await supabase.from("artistsongs").upsert(artistLinks, { onConflict: 'song_id, artist_id' });
-        if (linkErr) throw linkErr;
-
-        return { success: true };
-    } catch (error) {
-        console.error("Batch sync failed:", error);
-        throw error;
-    }
-}
-
-export async function addUpdateArtist(artist: SpotifyArtistDetailed) {
-    try {
-        // 1. Ensure the Artist row exists (but keep fully_loaded: false)
-        const { data: existingArtist } = await supabase
-            .from("artists")
-            .select("artist_id")
-            .eq("artist_id", artist.id)
-            .single();
-
-        if (!existingArtist) {
-            const { error: insError } = await supabase.from("artists").insert([{
-                artist_id: artist.id,
-                name: artist.name,
-                spotify_url: artist.external_urls.spotify,
-                fully_loaded: false, 
-                image_url: artist.images?.[0]?.url ?? null,
-            }]);
-            if (insError) throw insError;
-        }
-
-        // 2. Prepare and Upsert Albums
-        const { data: existingAlbums, error: fetchError } = await supabase
-            .from("albums")
-            .select("album_id, fully_loaded")
-            .in("album_id", artist.albums.map((a: any) => a.id));
-
-        if (fetchError) throw fetchError;
-
-        const existingMap = new Map(existingAlbums?.map(a => [a.album_id, a.fully_loaded]) ?? []);
-
-        const albumInserts = artist.albums.map((a: any) => ({
-            album_id: a.id,
-            title: a.name,
-            album_cover_url: a.images[0]?.url,
-            release_date: normalizeReleaseDate(a.release_date),
-            total_tracks: a.total_tracks,
-            spotify_url: a.external_urls.spotify,
-            fully_loaded: existingMap.get(a.id) === true, // Keep its status if it exists
-            album_type: a.album_type
-        }));
-
-        const { error: addingAlbums } = await supabase.from("albums").upsert(albumInserts);
-        if (addingAlbums) throw addingAlbums;
-
-        // 3. Link Artist and Albums
-        const albumArtistLinks = artist.albums.map((a: any) => ({ 
-            album_id: a.id, 
-            artist_id: artist.id 
-        }));
-
-        const { error: artistsalbumError } = await supabase
-            .from("artistalbum")
-            .upsert(albumArtistLinks);
-
-        if (artistsalbumError) throw artistsalbumError;
-
-        // 4. FINAL STEP: Mark as fully loaded
-        // We also update name and image here in case they changed on Spotify
-        const { error: finalUpdate } = await supabase
-            .from("artists")
-            .update({ 
-                name: artist.name,
-                image_url: artist.images?.[0]?.url ?? null,
-                fully_loaded: true 
-            })
-            .eq("artist_id", artist.id);
-
-        if (finalUpdate) throw finalUpdate;
-
-        return true;
-    } catch (error) {
-        console.error("Error adding artist to database:", error);
-        throw error;
-    }
-}
-
-
-// add album without adding all tracks
-export async function addAlbumSimple(album: SpotifyAlbum, artist_id: string) {
-    try {
-        // --- Check album ---
-        const { data: existingAlbum } = await supabase
-            .from("albums")
-            .select("*")
-            .eq("album_id", album.id)
-            .single();
-
-        if (!existingAlbum) {
-            const { error: albumError } = await supabase
-                .from("albums")
-                .insert([{
-                    album_id: album.id,
-                    title: album.name,
-                    spotify_url: album.external_urls.spotify,
-                    album_cover_url: album.images[0]?.url,
-                    release_date: normalizeReleaseDate(album.release_date),
-                    total_tracks: album.total_tracks,
-                    fully_loaded: false
-                }]);
-            if (albumError) throw albumError;
-        }
-
-        if (existingAlbum) {
-            return;
-        }
-
-
-
-        const { error: artistError } = await supabase
-            .from("artistalbum")
-            .upsert([{ album_id: album.id, artist_id: artist_id }]);
-
-        if (artistError) throw artistError
-
-
-    } catch (error) {
-        console.error("Error adding Album to database:", error);
-        throw error;
-    }
-}
-
-// add album with tracks 
-export async function addAlbumComplex(album: SpotifyAlbumWithTracks) {
-    try {
-
-        console.log(album)
-        // --- Check album ---
-        const { data: existingAlbum } = await supabase
-            .from("albums")
-            .select("*")
-            .eq("album_id", album.id)
-            .single();
-
-        if (!existingAlbum) {
-            const { error: albumError } = await supabase
-                .from("albums")
-                .insert([{
-                    album_id: album.id,
-                    title: album.name,
-                    spotify_url: album.external_urls.spotify,
-                    album_cover_url: album.images[0]?.url,
-                    release_date: normalizeReleaseDate(album.release_date),
-                    total_tracks: album.total_tracks,
-                    fully_loaded: true,
-                    album_type: album.album_type
-                }]);
-            if (albumError) throw albumError;
-        } else if (existingAlbum.fully_loaded === true) {
-            return; // Already loaded
-        } else {
-            const { error: albumUpdate } = await supabase
-                .from("albums")
-                .update([{
-                    title: album.name,
-                    spotify_url: album.external_urls.spotify,
-                    album_cover_url: album.images[0]?.url,
-                    release_date: normalizeReleaseDate(album.release_date),
-                    total_tracks: album.total_tracks,
-                    fully_loaded: true,
-                    album_type: album.album_type
-                }])
-                .eq("album_id", album.id);
-            if (albumUpdate) throw albumUpdate;
-        }
-
-        // --- Batch insert artists ---
-        const artistInserts = album.artists.map(a => ({
-            artist_id: a.id,
-            name: a.name,
-            spotify_url: a.external_urls.spotify
-        }));
-
-        await supabase
-            .from("artists")
-            .upsert(artistInserts);
-
-        // --- Link album & artists ---
-        const artistAlbumLinks = album.artists.map(a => ({
-            album_id: album.id,
-            artist_id: a.id
-        }));
-
-        await supabase
-            .from("artistalbum")
-            .upsert(artistAlbumLinks);
-
-        // --- Batch insert songs ---
-
-        console.log(album.tracks.items)
-        
-        const songInserts = album.tracks.items.map((s: SpotifyTrack) => ({
-            song_id: s.id,
-            title: s.name,
-            album_id: album.id,
-            spotify_url: s.external_urls.spotify,
-            pop: s.popularity, 
-            isrc: s.external_ids?.isrc || null
-        }));
-
-        await supabase
-            .from("songs")
-            .upsert(songInserts);
-
-        // --- Link songs & artists ---
-        const songArtistLinks = album.tracks.items.flatMap((s: any) =>
-            s.artists.map((a: any) => ({
-                song_id: s.id,
-                artist_id: a.id
-            }))
-        );
-
-        await supabase
-            .from("artistsongs")
-            .upsert(songArtistLinks);
-
-    } catch (error) {
-        console.error("Error adding Album to database:", error);
-        throw error;
-    }
-}
-
-
-
-
-export async function addListenedSong(songId: string, userId: string) {
-    const now = new Date();
-    const isoString = now.toISOString();
-    try {
-        const { error: addListen } = await supabase
-            .from("song_listens")
-            .insert({
-                song_id: songId,
-                user_id: userId,
-                listen_date: isoString,
-            });
-
-        if (addListen) throw addListen;
-
-    } catch (error) {
-        console.log(error);
-    }
-}
-
-export async function addListenedAlbum(albumId: string, userId: string) {
-    const now = new Date();
-    const isoString = now.toISOString();
-    try {
-        const { error: addListen } = await supabase
-            .from("album_listens")
-            .insert({
-                album_id: albumId,
-                user_id: userId,
-                listen_date: isoString,
-            });
-
-        if (addListen) throw addListen;
-
-    } catch (error) {
-        console.log(error);
-    }
-}
-
-export async function hasListenedSong(userId: string, songId: string): Promise<Boolean> {
-    try {
-
-        const { data: listened } = await supabase
-            .from("song_listens")
-            .select("*")
-            .eq("song_id", songId)
-            .eq("user_id", userId)
-            .single();
-
-        if (!listened) return false;
-
-        return true;
-    } catch (error) {
-        console.error('Failed to fetch listened:', error);
-        return false;
-    }
-}
-
-export async function hasListenedAlbum(userId: string, albumId: string): Promise<Boolean> {
-    try {
-
-        const { data: listened } = await supabase
-            .from("album_listens")
-            .select("*")
-            .eq("album_id", albumId)
-            .eq("user_id", userId)
-            .single();
-
-        if (!listened) return false;
-
-        return true;
-    } catch (error) {
-        console.error('Failed to fetch listened:', error);
-        return false;
-    }
-}
-
-export async function removeListenedSong(songId: string, userId: string) {
-    try {
-        await supabase
-            .from("song_listens")
-            .delete()
-            .eq("song_id", songId)
-            .eq("user_id", userId);
-
-    } catch (error) {
-        console.error('Failed to delete listen:', error);
-        return false;
-    }
-}
-
-
-export async function removeListenedAlbum(albumId: string, userId: string) {
-    try {
-        await supabase
-            .from("album_listens")
-            .delete()
-            .eq("album_id", albumId)
-            .eq("user_id", userId);
-
-    } catch (error) {
-        console.error('Failed to delete listen:', error);
-        return false;
-    }
-}
-
-
-export async function addUpdateRatingSong(songId: string, userId: string, rating: number) {
-    const now = new Date();
-    const isoString = now.toISOString();
-    try {
-        const { error: addListen } = await supabase
-            .from("song_rating")
-            .upsert({
-                song_id: songId,
-                user_id: userId,
-                rating_date: isoString,
-                rating: rating
-            })
-
-        const { data: getListen, error: getListenError } = await supabase
-            .from("song_listens")
-            .select("*")
-            .eq("song_id", songId)
-            .eq("user_id", userId)
-
-        if (getListenError) throw getListenError;
-
-        if (getListen.length === 0) {
-            const { error: listenError } = await supabase
-                .from("song_listens")
-                .insert({ song_id: songId, user_id: userId, listen_date: isoString })
-            if (listenError) throw listenError;
-        }
-
-        if (addListen) throw addListen;
-
-    } catch (error) {
-        console.log(error);
-    }
-}
-
-export async function deleteRaitingSong(songId: string, userId: string) {
-    try {
-        const { error: deleteRating } = await supabase
-            .from("song_rating")
-            .delete()
-            .eq("user_id", userId)
-            .eq("song_id", songId);
-
-        if (deleteRating) throw deleteRating;
-
-    } catch (error) {
-        console.log(error);
-    }
-}
-
-export async function getRatingSong(songId: string, userId: string): Promise<number> {
-    try {
-        const { data: rating } = await supabase
-            .from("song_rating")
-            .select("*")
-            .eq("song_id", songId)
-            .eq("user_id", userId)
-            .single();
-
-        if (!rating) return 0;
-
-
-        return rating.rating;
-    } catch (error) {
-        console.error('Failed to fetch raiting:', error);
-        return 0;
-    }
-}
-
-export async function getAllRatingsOfaSong(songId: string): Promise<{
-    counts: { rating: number; count: number }[];
-    average: number;
-    total: number;
-}> {
-    try {
-        const { data: ratings, error } = await supabase
-            .from("song_rating")
-            .select("rating")
-            .eq("song_id", songId);
-
-        if (error) throw error;
-        if (!ratings || ratings.length === 0) return { counts: [], average: 0, total: 0 };
-
-        // 1. Initialize all 10 Letterboxd-style buckets (0.5 to 5.0)
-        const possibleRatings = [0.5, 1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5];
-        const countsMap = new Map(possibleRatings.map(r => [r, 0]));
-
-        let sum = 0;
-        const total = ratings.length;
-
-        ratings.forEach((r) => {
-            // Postgres numeric types often return as strings, so cast to Number
-            const val = Number(r.rating); 
-            sum += val;
-
-            // Increment the specific bucket
-            if (countsMap.has(val)) {
-                countsMap.set(val, (countsMap.get(val) || 0) + 1);
-            }
-        });
-
-        // 2. Format for Recharts
-        const counts = possibleRatings.map((r) => ({
-            rating: r,
-            count: countsMap.get(r) || 0,
-        }));
-
-        // 3. Calculate accurate decimal average
-        const average = sum / total;
-
-        return { counts, average, total };
-    } catch (error) {
-        console.error("Failed to fetch song ratings:", error);
-        return { counts: [], average: 0, total: 0 };
-    }
-}
-
-
-
-export async function addUpdateRatingAlbum(albumId: string, userId: string, rating: number) {
-    const now = new Date();
-    const isoString = now.toISOString();
-    try {
-        const { error: addListen } = await supabase
-            .from("album_rating")
-            .upsert({
-                album_id: albumId,
-                user_id: userId,
-                rating_date: isoString,
-                rating: rating
-            })
-
-        const { data: updateListen, error: listenError } = await supabase
-            .from("album_listens")
-            .upsert({ album_id: albumId, user_id: userId, listen_date: isoString })
-
-        if (listenError) throw listenError;
-        if (addListen) throw addListen;
-
-        if (!updateListen) throw new Error("Failed to update listen time");
-
-    } catch (error) {
-        console.log(error);
-    }
-}
-
-export async function deleteRaitingAlbum(albumId: string, userId: string) {
-    try {
-        const { error: deleteRating } = await supabase
-            .from("album_rating")
-            .delete()
-            .eq("user_id", userId)
-            .eq("album_id", albumId);
-
-        if (deleteRating) throw deleteRating;
-
-    } catch (error) {
-        console.log(error);
-    }
-}
-
-
-export async function getRatingAlbum(albumId: string, userId: string): Promise<number> {
-    try {
-        const { data: rating } = await supabase
-            .from("album_rating")
-            .select("*")
-            .eq("album_id", albumId)
-            .eq("user_id", userId)
-            .single();
-
-        if (!rating) return 0;
-
-
-        return rating.rating;
-    } catch (error) {
-        console.error('Failed to fetch raiting:', error);
-        return 0;
-    }
-}
-
-
-export async function getAllRatingsOfAlbum(albumId: string): Promise<{
-    counts: { rating: number; count: number }[];
-    average: number;
-    total: number;
-}> {
-    try {
-        const { data: ratings, error } = await supabase
-            .from("album_rating")
-            .select("rating")
-            .eq("album_id", albumId);
-
-        if (error) throw error;
-        if (!ratings || ratings.length === 0) return { counts: [], average: 0, total: 0 };
-
-        // 1. Initialize all 10 Letterboxd-style buckets (0.5 to 5.0)
-        const possibleRatings = [0.5, 1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5];
-        const countsMap = new Map(possibleRatings.map(r => [r, 0]));
-
-        let sum = 0;
-        const total = ratings.length;
-
-        ratings.forEach((r) => {
-            const val = Number(r.rating); 
-            sum += val;
-
-            // Increment the specific half-star bucket
-            if (countsMap.has(val)) {
-                countsMap.set(val, (countsMap.get(val) || 0) + 1);
-            }
-        });
-
-        // 2. Format for Recharts
-        const counts = possibleRatings.map((r) => ({
-            rating: r,
-            count: countsMap.get(r) || 0,
-        }));
-
-        const average = sum / total;
-
-        return { counts, average, total };
-    } catch (error) {
-        console.error("Failed to fetch ratings:", error);
-        return { counts: [], average: 0, total: 0 };
-    }
-}
-
-
-export async function getListenedWithLimit(
-    userId: string, 
-    limit: number, 
-    offset: number,
-    sortBy: "newest" | "oldest" = "newest",
-    filterType: "all" | "song" | "album" = "all",
-    searchQuery: string = ""
-): Promise<Listened[]> {
-    try {
-        // 1. Simple query to the enhanced view
-        let query = supabase
-            .from("user_history_view")
-            .select("*")
-            .eq("user_id", userId);
-
-        // 2. Optional: Search by name
-        if (searchQuery) {
-            query = query.ilike("name", `%${searchQuery}%`);
-        }
-
-        // 3. Media Type Filtering
-        if (filterType !== "all") {
-            query = query.eq("type", filterType);
-        }
-
-        // 4. Chronological Ordering
-        query = query.order("listen_date", { ascending: sortBy === "oldest" });
-
-        // 5. Server-side Pagination
-        const { data: history, error } = await query.range(offset, offset + limit - 1);
-
-        if (error) throw error;
-
-        // 6. Instant Mapping
-        return (history || []).map((h: any) => ({
-            type: h.type,
-            id: h.item_id,
-            name: h.name ?? "Unknown",
-            album_cover_url: h.album_cover_url,
-            listen_date: h.listen_date,
-            createdAt: h.listen_date 
-        }));
-
-    } catch (error) {
-        console.error("Failed to fetch listened history:", error);
-        return [];
-    }
-}
-
-// Fetch all (no limit)
-export async function getListened(userId: string): Promise<Listened[]> {
-    try {
-        const { data: songsListened, error: songsError } = await supabase
-            .from("song_listens")
-            .select(`
-        listen_date,
-        song_id,
-        songs (
-          title,
-          albums:songs_album_id_fkey (
-            album_cover_url
-          )
-        )
-      `)
-            .eq("user_id", userId);
-
-        if (songsError) throw songsError;
-
-        const { data: albumsListened, error: albumsError } = await supabase
-            .from("album_listens")
-            .select(`
-        listen_date,
-        album_id,
-        albums (
-          title,
-          album_cover_url
-        )
-      `)
-            .eq("user_id", userId);
-
-        if (albumsError) throw albumsError;
-
-        const combined: Listened[] = [
-            ...(songsListened?.map((s: any) => ({
-                type: "song" as const,
-                id: s.song_id,
-                name: s.songs?.title ?? "Unknown Song",
-                album_cover_url: s.songs?.albums?.album_cover_url ?? null,
-                listen_date: s.listen_date,
-            })) ?? []),
-            ...(albumsListened?.map((a: any) => ({
-                type: "album" as const,
-                id: a.album_id,
-                name: a.albums?.title ?? "Unknown Album",
-                album_cover_url: a.albums?.album_cover_url ?? null,
-                listen_date: a.listen_date,
-            })) ?? []),
-        ];
-
-        return combined.sort(
-            (a, b) =>
-                new Date(b.listen_date).getTime() - new Date(a.listen_date).getTime()
-        );
-    } catch (error) {
-        console.error("Failed to fetch listened:", error);
-        return [];
-    }
-}
-
-export async function getRatedWithLimit(
-    userId: string, 
-    limit: number, 
-    offset: number,
-    sortBy: "newest" | "oldest" | "rating" = "newest",
-    filterType: "all" | "song" | "album" = "all",
-    searchQuery: string = ""
-): Promise<RatingGeneral[]> {
-    try {
-        let query = supabase
-            .from("user_ratings_view")
-            .select("*") // Everything is already flattened in the view
-            .eq("user_id", userId);
-
-        // 1. Search Logic
-        if (searchQuery) {
-            query = query.ilike("title", `%${searchQuery}%`);
-        }
-
-        // 2. Media Filter
-        if (filterType !== "all") {
-            query = query.eq("type", filterType);
-        }
-
-        // 3. Ordering
-        if (sortBy === "oldest") {
-            query = query.order("rating_date", { ascending: true });
-        } else if (sortBy === "rating") {
-            query = query.order("rating", { ascending: false });
-        } else {
-            query = query.order("rating_date", { ascending: false });
-        }
-
-        const { data: ratings, error } = await query.range(offset, offset + limit - 1);
-
-        if (error) throw error;
-
-        // No more async maps! Just return the data.
-        return (ratings || []).map((r: any) => ({
-            type: r.type,
-            id: r.item_id,
-            rating: r.rating,
-            title: r.title ?? "Unknown",
-            album_cover_url: r.album_cover_url,
-            rating_date: r.rating_date,
-            createdAt: r.rating_date
-        }));
-
-    } catch (error) {
-        console.error('Failed to fetch ratings:', error);
-        return [];
-    }
-}
-
-export async function getRated(userId: string): Promise<RatingGeneral[]> {
-    try {
-        const { data: songsRated, error: songsError } = await supabase
-            .from("song_rating")
-            .select(`
-                rating,
-                rating_date,
-                song_id,
-                songs (
-                title,
-                albums:songs_album_id_fkey (
-                    album_cover_url
-                )
-                )
-            `)
-            .eq("user_id", userId);
-
-        if (songsError) throw songsError;
-
-        const { data: albumsRated, error: albumsError } = await supabase
-            .from("album_rating")
-            .select(`
-                rating_date,
-                rating,
-                album_id,
-                albums (
-                    title,
-                    album_cover_url
-                )
-            `)
-            .eq("user_id", userId);
-
-        if (albumsError) throw albumsError;
-
-        const combined: RatingGeneral[] = [
-            ...(songsRated.map((s: any) => ({
-                type: "song" as const,
-                id: s.song_id,
-                rating: s.rating,
-                title: s.songs?.title ?? "Unknown Song",
-                album_cover_url: s.songs?.albums?.album_cover_url ?? null,
-                rating_date: s.rating_date,
-            })) ?? []),
-            ...(albumsRated?.map((a: any) => ({
-                type: "album" as const,
-                id: a.album_id,
-                rating: a.rating,
-                title: a.albums?.title ?? "Unknown Album",
-                album_cover_url: a.albums?.album_cover_url ?? null,
-                rating_date: a.rating_date,
-            })) ?? []),
-        ];
-
-
-        return combined.sort(
-            (a, b) =>
-                new Date(b.rating_date).getTime() - new Date(a.rating_date).getTime()
-        );
-    } catch (error) {
-        console.error('Failed to fetch user:', error);
-        return [];
-    }
-}
-
-export async function getReviewedWithLimit(
-    userId: string, 
-    limit: number, 
-    offset: number,
-    sortBy: "newest" | "oldest" | "rating" = "newest", 
-    filterType: "all" | "song" | "album" = "all" ,
-    searchQuery: string = ""    
-): Promise<Review[]> {
-    try {
-        // 1. Query the VIEW instead of the TABLE
-        let query = supabase
-            .from("searchable_reviews")
-            .select(`
-                *,
-                reviewlikes(count)
-            `)
-            .eq("user_id", userId);
-
-        // 2. Search logic (Now this works perfectly because everything is in one table)
-        if (searchQuery) {
-            query = query.or(`review_text.ilike.%${searchQuery}%, song_title.ilike.%${searchQuery}%, album_title.ilike.%${searchQuery}%, review_title.ilike.%${searchQuery}%`);
-        }
-
-        // 3. Filtering
-        if (filterType === "song") query = query.not("song_id", "is", null);
-        if (filterType === "album") query = query.not("album_id", "is", null);
-
-        // 4. Ordering
-        query = query.order(sortBy === "oldest" ? "created_at" : "created_at", { ascending: sortBy === "oldest" });
-
-        const { data: reviews, error } = await query.range(offset, offset + limit - 1);
-
-        if (error) throw error;
-
-        return reviews.map((r: any) => ({
-            reviewId: r.review_id,
-            text: r.review_text,
-            id: r.song_id ?? r.album_id,
-            name: r.song_title ?? r.album_title ?? "Unknown",
-            userId: r.user_id,
-            type: r.song_id ? "song" : "album",
-            createdAt: r.created_at,
-            title: r.review_title,
-            rating: r.rating,
-            album_cover_url: r.effective_cover_url ?? "",
-            likes: r.reviewlikes?.[0]?.count || 0,
-        }));
-    } catch (err) {
-        console.error("Failed to fetch user reviews:", err);
-        return [];
-    }
-}
-
-export async function getReviewed(userId: string): Promise<Review[]> {
-    try {
-        const { data: reviews, error } = await supabase
-            .from("reviews")
-            .select(`
-                review_id,
-                review_text,
-                review_title,
-                song_id,
-                album_id,
-                created_at,
-                user_id
-            `)
-            .eq("user_id", userId)
-            .order("created_at", { ascending: false });
-        if (error) throw error;
-        if (!reviews) return [];
-
-
-
-
-        const results = await Promise.all(
-            reviews.map(async (r: any) => {
-                let coverUrl = "";
-                const reviewType: "song" | "album" = r.song_id ? "song" : "album";
-                let name = "";
-
-                if (r.album_id) {
-                    const { data: album } = await supabase
-                        .from("albums")
-                        .select("*")
-                        .eq("album_id", r.album_id)
-                        .single();
-                    coverUrl = album?.album_cover_url ?? "";
-                    name = album?.title ?? "";
-                } else if (r.song_id) {
-                    const { data: song } = await supabase
-                        .from("songs")
-                        .select("*, album:albums(album_cover_url)")
-                        .eq("song_id", r.song_id)
-                        .single();
-                    coverUrl = song?.album.album_cover_url ?? "";
-                    name = song?.title ?? "";
-                }
-
-                return {
-                    reviewId: r.review_id,
-                    text: r.review_text,
-                    id: r.song_id ?? r.album_id,
-                    title: r.review_title,
-                    name: name,
-                    userId: r.user_id,
-                    type: reviewType,
-                    createdAt: r.created_at,
-                    album_cover_url: coverUrl,
-                    likes: 0
-                };
-            })
-        );
-        return results;
-    } catch (err) {
-        console.error("Failed to fetch user reviews:", err);
-        return [];
-    }
-}
-
-
-
-
-
-export async function getRecentFollowedActivities(
-    userId: string,
-    limit: number = 10,
-    offset: number = 0,
-): Promise<Activity[]> {
-    try {
-        // 1️⃣ Get list of followed user IDs
-        const { data: followed, error: followsError } = await supabase
-            .from("followers")
-            .select("following_id")
-            .eq("follower_id", userId);
-
-        if (followsError) throw followsError;
-        if (!followed || followed.length === 0) return [];
-
-        const followedIds = followed.map((f) => f.following_id);
-
-        const poolSize = limit * 3;
-
-        const [
-            { data: songListens, error: songListensError },
-            { data: albumListens, error: albumListensError },
-            { data: reviews, error: reviewsError },
-            { data: songRatings, error: songRatingsError },
-            { data: albumRatings, error: albumRatingsError }
-        ] = await Promise.all([
-            supabase
-                .from("song_listens")
-                .select(`
-                    user_id,
-                    song_id,
-                    listen_date,
-                    song:songs(title, album_id, albums(album_cover_url, title)),
-                    user:users(username)
-                    `)
-                .in("user_id", followedIds)
-                .order("listen_date", { ascending: false })
-                .limit(poolSize),
-            supabase
-                .from("album_listens")
-                .select(`
-                    user_id,
-                    album_id,
-                    listen_date,
-                    album:albums(title, album_cover_url),
-                    user:users(username)
-                    `)
-                .in("user_id", followedIds)
-                .order("listen_date", { ascending: false })
-                .limit(poolSize),
-            supabase
-                .from("reviews")
-                .select(`
-                    review_id,
-                    user_id,
-                    review_text,
-                    review_type,
-                    song_id,
-                    album_id,
-                    created_at,
-                    song:songs(title, albums(album_cover_url)),
-                    album:albums(title, album_cover_url),
-                    user:users(username)
-                    `)
-                .in("user_id", followedIds)
-                .order("created_at", { ascending: false })
-                .limit(poolSize),
-            supabase
-                .from("song_rating")
-                .select(`
-                    user_id,
-                    rating,
-                    song_id,
-                    rating_date,
-                    song:songs(title, album_id, albums(album_cover_url, title)),
-                    user:users(username)
-                `)
-                .in("user_id", followedIds)
-                .order("rating_date", { ascending: false })
-                .limit(poolSize),
-            supabase
-                .from("album_rating")
-                .select(`
-                    user_id,
-                    rating,
-                    album_id,
-                    rating_date,
-                    album:albums(title, album_cover_url),
-                    user:users(username)
-                    `)
-                .in("user_id", followedIds)
-                .order("rating_date", { ascending: false })
-                .limit(poolSize),
-        ]);
-
-        if (songListensError) throw songListensError;
-        if (albumListensError) throw albumListensError;
-        if (reviewsError) throw reviewsError;
-        if (songRatingsError) throw songRatingsError;
-        if (albumRatingsError) throw albumRatingsError;
-
-        const activities: Activity[] = [];
-
-        // helper to handle supabase returning nested relations as arrays
-        const first = (v: any) => (Array.isArray(v) ? v[0] : v) as any;
-
-        // SONG LISTENS
-        for (const l of songListens || []) {
-            const userObj = first(l.user);
-            const songObj = first(l.song);
-            const songAlbum = first(songObj?.albums);
-            activities.push({
-                id: `songlisten-${l.user_id}-${l.song_id}-${l.listen_date}`,
-                userId: l.user_id,
-                username: userObj?.username ?? "Unknown",
-                type: "listen",
-                targetType: "song",
-                targetId: l.song_id,
-                targetName: songObj?.title ?? "Unknown Song",
-                album_cover_url: songAlbum?.album_cover_url ?? "",
-                date: l.listen_date,
-                targetAlbumId: songObj.album_id,
-                targetAlbumName : songAlbum?.title ?? "Unknown Album",
-                isAggregated: false,
-                groupedActivities: []
-            });
-        }
-
-        // ALBUM LISTENS
-        for (const l of albumListens || []) {
-            const userObj = first(l.user);
-            const albumObj = first(l.album);
-            activities.push({
-                id: `albumlisten-${l.user_id}-${l.album_id}-${l.listen_date}`,
-                userId: l.user_id,
-                username: userObj?.username ?? "Unknown",
-                type: "listen",
-                targetType: "album",
-                targetId: l.album_id,
-                targetName: albumObj?.title ?? "Unknown Album",
-                album_cover_url: albumObj?.album_cover_url ?? "",
-                date: l.listen_date,
-                targetAlbumId: l.album_id,
-                targetAlbumName : albumObj?.title ?? "Unknown Album",
-                isAggregated: false,
-                groupedActivities: []
-            });
-        }
-
-        // REVIEWS
-        for (const r of reviews || []) {
-            const userObj = first(r.user);
-            const songObj = first(r.song);
-            const albumObj = first(r.album);
-            const songAlbum = first(songObj?.albums);
-            activities.push({
-                id: `${r.review_id}`,
-                userId: r.user_id,
-                username: userObj?.username ?? "Unknown",
-                type: "review",
-                targetType: r.review_type === "song" ? "song" : "album",
-                targetId: r.review_type === "song" ? r.song_id : r.album_id,
-                targetName:
-                    r.review_type === "song"
-                        ? songObj?.title ?? "Unknown Song"
-                        : albumObj?.title ?? "Unknown Album",
-                album_cover_url:
-                    r.review_type === "song"
-                        ? songAlbum?.album_cover_url ?? ""
-                        : albumObj?.album_cover_url ?? "",
-                date: r.created_at,
-                text: r.review_text,
-                targetAlbumId: r.album_id,
-                targetAlbumName : songAlbum?.title ?? "Unknown Album",
-                isAggregated: false,
-                groupedActivities: []
-            });
-        }
-
-        // SONG RATINGS
-        for (const r of songRatings || []) {
-            const userObj = first(r.user);
-            const songObj = first(r.song);
-            const songAlbum = first(songObj?.albums);
-            activities.push({
-                id: `songrating-${r.user_id}-${r.song_id}-${r.rating_date}`,
-                userId: r.user_id,
-                username: userObj?.username ?? "Unknown",
-                type: "rating",
-                targetType: "song",
-                targetId: r.song_id,
-                targetName: songObj?.title ?? "Unknown Song",
-                album_cover_url: songAlbum?.album_cover_url ?? "",
-                date: r.rating_date,
-                rating: r.rating,
-                targetAlbumId:  songObj.album_id,
-                targetAlbumName : songAlbum?.title ?? "Unknown Album",
-                isAggregated: false,
-                groupedActivities: []
-            });
-        }
-
-        // ALBUM RATINGS
-        for (const r of albumRatings || []) {
-            const userObj = first(r.user);
-            const albumObj = first(r.album);
-            activities.push({
-                id: `albumrating-${r.user_id}-${r.album_id}-${r.rating_date}`,
-                userId: r.user_id,
-                username: userObj?.username ?? "Unknown",
-                type: "rating",
-                targetType: "album",
-                targetId: r.album_id,
-                targetName: albumObj?.title ?? "Unknown Album",
-                album_cover_url: albumObj?.album_cover_url ?? "",
-                date: r.rating_date,
-                rating: r.rating,
-                targetAlbumId: r.album_id,
-                targetAlbumName : albumObj?.title ?? "Unknown Album",
-                isAggregated: false,
-                groupedActivities: []
-            });
-        }
-
-        activities.sort(
-            (a, b) => new Date(b.date).getTime() - new Date(a.date).getTime()
-        );
-
-        const aggregatedActivities = combineAndAggregate(activities);
-
-        console.log(aggregatedActivities);
-
-        const paginated = aggregatedActivities.slice(offset, offset + limit);
-
-        const uniqueUserIds = Array.from(new Set(paginated.map(a => a.userId)));
-
-        const avatarPromises = await Promise.all(
-            uniqueUserIds.map(async (id) => ({
-                id,
-                avatarUrl: await getProfileUrl(id),
-            }))
-        );
-        const avatarMap = Object.fromEntries(avatarPromises.map((a) => [a.id, a.avatarUrl]));
-
-        return paginated.map((a) => ({
-            ...a,
-            profileUrl: avatarMap[a.userId] || "/assets/default-avatar.png",
-        }));
-
-
-    } catch (err) {
-        console.error("Failed to fetch recent followed activities:", err);
-        return [];
-    }
-}
-function combineAndAggregate(activities: Activity[]): Activity[] {
-    if (!activities.length) return [];
-
-    const groups = new Map<string, Activity[]>();
-    const TIME_WINDOW_MS = 15 * 60 * 1000; // Increased to 15m for better grouping
-
-    // 1. Initial Grouping by User + Album + Time Window
-    activities.forEach(activity => {
-        const albumId = activity.targetType === "album" ? activity.targetId : activity.targetAlbumId;
-        if (!albumId) {
-            groups.set(`single-${activity.id}`, [activity]);
-            return;
-        }
-
-        const albumKeyPrefix = `${activity.userId}-${albumId}`;
-        let foundGroupKey: string | null = null;
-
-        for (const [key, group] of groups.entries()) {
-            if (key.startsWith(albumKeyPrefix)) {
-                const groupStartTime = new Date(group[0].date).getTime();
-                const activityTime = new Date(activity.date).getTime();
-                if (Math.abs(activityTime - groupStartTime) <= TIME_WINDOW_MS) {
-                    foundGroupKey = key;
-                    break;
-                }
-            }
-        }
-
-        if (foundGroupKey) {
-            groups.get(foundGroupKey)!.push(activity);
-        } else {
-            const newKey = `${albumKeyPrefix}-${new Date(activity.date).getTime()}`;
-            groups.set(newKey, [activity]);
-        }
-    });
-
-    const newFeed: Activity[] = [];
-
-    for (const group of groups.values()) {
-        // --- SECONDARY AGGREGATION: Merge Listen + Rating for the same song ---
-        const mergedByTarget = new Map<string, Activity>();
-        
-        group.forEach(act => {
-            const key = `${act.targetType}-${act.targetId}`;
-            if (!mergedByTarget.has(key)) {
-                mergedByTarget.set(key, { ...act });
-            } else {
-                const existing = mergedByTarget.get(key)!;
-                // Merge Rating into Listen (or vice versa)
-                if (act.rating) existing.rating = act.rating;
-                if (act.text) existing.text = act.text;
-                // Keep the most recent date
-                if (new Date(act.date) > new Date(existing.date)) {
-                    existing.date = act.date;
-                }
-                // If one was a 'review' or 'rating', upgrade the type from 'listen'
-                if (act.type === "review" || act.type === "rating") {
-                    existing.type = act.type;
-                }
-            }
-        });
-
-        const uniqueActivitiesInGroup = Array.from(mergedByTarget.values());
-
-        // 2. Decide how to display this group
-        if (uniqueActivitiesInGroup.length === 1) {
-            newFeed.push(uniqueActivitiesInGroup[0]);
-        } else {
-            // Logic for Aggregated "Juked" Post
-            const explicitAlbumActivity = 
-                uniqueActivitiesInGroup.find(a => a.targetType === "album" && a.type === "review") || 
-                uniqueActivitiesInGroup.find(a => a.targetType === "album" && a.type === "rating") || 
-                uniqueActivitiesInGroup.find(a => a.targetType === "album");
-
-            const representative = uniqueActivitiesInGroup[0];
-            
-            let headline: Activity;
-            let children: Activity[];
-
-            if (explicitAlbumActivity) {
-                headline = explicitAlbumActivity;
-                children = uniqueActivitiesInGroup.filter(a => a.id !== explicitAlbumActivity.id);
-            } else {
-                // Synthetic Headline
-                headline = {
-                    ...representative,
-                    targetType: "album",
-                    targetId: representative.targetAlbumId!,
-                    targetName: representative.targetAlbumName!,
-                    type: "grouped",
-                    rating: undefined,
-                    text: undefined,
-                };
-                children = uniqueActivitiesInGroup;
-            }
-
-            newFeed.push({
-                ...headline,
-                id: `aggregated-${headline.userId}-${headline.targetId}-${new Date(headline.date).getTime()}`,
-                isAggregated: true,
-                groupedActivities: children,
-                type: "grouped",
-                targetName: headline.targetAlbumName || headline.targetName,
-            });
-        }
-    }
-
-    return newFeed.sort((a, b) => new Date(b.date).getTime() - new Date(a.date).getTime());
-}
-export async function getTrendingData(limit: number = 10): Promise<TrendingResponse> {
-    try {
-        const [songsRes, albumsRes, reviewsRes] = await Promise.all([
-            supabase.from("trending_songs").select("*").order("play_count", { ascending: false }).limit(limit),
-            supabase.from("trending_albums").select("*").order("play_count", { ascending: false }).limit(limit),
-            supabase.from("trending_reviews").select("*").order("like_count", { ascending: false }).limit(limit)
-        ]);
-
-        return {
-            topSongs: (songsRes.data || []).map(s => ({
-                songId: s.song_id,
-                title: s.song_title,
-                albumCoverUrl: s.album_cover_url,
-                playCount: s.play_count
-            })),
-            topAlbums: (albumsRes.data || []).map(a => ({
-                albumId: a.album_id,
-                title: a.album_title,
-                albumCoverUrl: a.album_cover_url,
-                playCount: a.play_count
-            })),
-            recentReviews: (reviewsRes.data || []).map(r => ({
-                reviewId: r.review_id,
-                userId: r.user_id,
-                username: r.username,
-                targetType: r.review_type as "song" | "album",
-                targetId: r.review_type === "song" ? r.song_id : r.album_id,
-                targetName: r.target_name,
-                albumCoverUrl: r.album_cover_url,
-                reviewText: r.review_text,
-                createdAt: r.created_at,
-                likeCount: r.like_count || 0,
-            }))
-        };
-    } catch (err) {
-        console.error(err);
-        return { topSongs: [], topAlbums: [], recentReviews: [] };
-    }
-}
-
-export async function getTrendingSongs( limit: number = 10) : Promise<SongActivity[]> {
-    try {
-        const songRes = await supabase.from("trending_songs").select("*").order("play_count", { ascending: false }).limit(limit);
-
-        return (songRes.data || []).map(s => ({
-            songId: s.song_id,
-            title: s.song_title,
-            albumCoverUrl: s.album_cover_url,
-            playCount: s.play_count
-        }));
-    }
-    catch (err) {
-        console.error(err);
-        return [];
-    }
-}
-
-export async function getTrendingAlbums( limit: number = 10) : Promise<AlbumActivity[]> {
-    try {
-        const albumsRes = await supabase.from("trending_albums").select("*").order("play_count", { ascending: false }).limit(limit);
-
-        return (albumsRes.data || []).map(a => ({
-            albumId: a.album_id,
-            title: a.album_title,
-            albumCoverUrl: a.album_cover_url,
-            playCount: a.play_count
-        }));
-    }
-    catch (err) {
-        console.error(err);
-        return [];
-    }
-}
-
-export async function getTrendingReviews( limit: number = 10) : Promise<ReviewActivity[]> {
-    try {
-        const reviewsRes = await supabase.from("trending_reviews").select("*").order("like_count", { ascending: false }).limit(limit);
-
-        return  (reviewsRes.data || []).map(r => ({
-            reviewId: r.review_id,
-            userId: r.user_id,
-            username: r.username,
-            targetType: r.review_type as "song" | "album",
-            targetId: r.review_type === "song" ? r.song_id : r.album_id,
-            targetName: r.target_name,
-            albumCoverUrl: r.album_cover_url,
-            reviewText: r.review_text,
-            createdAt: r.created_at,
-            likeCount: r.like_count || 0,
-        }));
-    }
-    catch (err) {
-        console.error(err);
-        return [];
-    }
-}
-
-
-export async function getFollowerSuggestions(userId: string) {
-    if (!userId) return [];
-
-    const { data: suggestions, error: sugError } = await supabase
-        .rpc('get_top_mutual_suggestions', { target_user_id: userId });
-
-    if (sugError) throw sugError;
-    if (!suggestions || suggestions.length === 0) return [];
-
-    const ids = suggestions.map((s: any) => s.suggested_user);
-
-    // Fetch profiles in one request
-    const { data: profiles, error: profError } = await supabase
-        .from('users')
-        .select('user_id, username, name')
-        .in('user_id', ids);
-
-    if (profError) throw profError;
-
-    const avatarPromises = await Promise.all(
-        ids.map(async (id: string) => ({
-            id,
-            avatar_url: await getProfileUrl(id),
-        }))
-    );
-    const avatarMap = Object.fromEntries(avatarPromises.map((a) => [a.id, a.avatar_url]));
-
-    const profilesWithAvatars = profiles.map((a) => ({
-        ...a,
-        id: a.user_id,
-        avatar_url: avatarMap[a.user_id] || "/assets/default-avatar.png",
-    }));
-
-    // merge counts with profiles
-    return suggestions.map((s: any) => ({
-        mutual_count: s.mutual_count,
-        ...profilesWithAvatars.find(p => p.user_id === s.suggested_user)
-    }));
-}
-
-
-// export const getPlaylistById = async (playlistId: string): Promise<Playlist | null> => {
-//     try {
-//         const { data, error } = await supabase
-//             .from('playlists')
-//             .select(`
-//                 playlist_id,
-//                 name,
-//                 description,
-//                 created_at,
-//                 cover_url,
-//                 playlist_creators (
-//                     user:users (*) 
-//                 ),
-//                 playlist_items (
-//                     id,
-//                     order,
-//                     type,
-//                     song:songs (
-//                         album:albums(album_cover_url, title), 
-//                         title, 
-//                         song_id, 
-//                         album_id, 
-//                         isrc, 
-//                         artists:artists (*)
-//                     ),
-//                     album:albums (
-//                         album_id,
-//                         title,
-//                         album_cover_url,
-//                         artists:artists (*),
-//                         songs:songs (
-//                             song_id,
-//                             title,
-//                             isrc,
-//                             artists:artists (*)
-//                         )
-//                     )
-//                 ),
-//                 songCount:playlist_items(count)
-//             `)
-//             .eq('playlist_id', playlistId)
-//             .order('order', { referencedTable: 'playlist_items', ascending: true })
-//             .single();
-
-//         if (error || !data) throw error || new Error("Playlist not found");
-
-//         // 1. Resolve Creators (Existing logic)
-//         const resolvedCreators = await Promise.all(
-//             data.playlist_creators.map(async (pc: any) => ({
-//                 ...pc.user,
-//                 imageUrl: await getProfileUrl(pc.user.user_id), 
-//                 accountId: pc.user.user_id,
-//                 name: pc.user.name
-//             }))
-//         );
-
-//         // 2. Prepare Inputs for Enrichment (Extract tracks from both songs and albums)
-//         const trackInputs: any[] = [];
-//         data.playlist_items.forEach((item: any) => {
-//             if (item.type === 'song' && item.song) {
-//                 trackInputs.push({
-//                     songId: item.song.song_id,
-//                     title: item.song.title,
-//                     artist: item.song.artists.map((a: any) => a.name).join(", "),
-//                     isrc: item.song.isrc
-//                 });
-//             } else if (item.type === 'album' && item.album) {
-//                 item.album.songs.forEach((as: any) => {
-//                     trackInputs.push({
-//                         songId: as.song_id,
-//                         title: as.title,
-//                         artist: as.artists.map((a: any) => a.name).join(", "),
-//                         isrc: as.isrc
-//                     });
-//                 });
-//             }
-//         });
-
-//         // 3. Invoke Edge Function for Previews
-//         const { data: enrichmentData } = await supabase.functions.invoke('enrich-album', {
-//             body: { tracks: trackInputs }
-//         });
-
-//         const previewMap = new Map();
-//         enrichmentData?.tracks?.forEach((t: any) => previewMap.set(t.songId || t.song_id, t.preview_url));
-
-//         return {
-//             playlistId: data.playlist_id,
-//             name: data.name,
-//             description: data.description,
-//             createdAt: data.created_at,
-//             coverUrl: data.cover_url,
-//             creators: resolvedCreators,
-//             songCount: data.songCount?.[0]?.count || 0,
-//             albumCount: 0, 
-//             totalTracks: 0,
-//             items: data.playlist_items.map((item: any) => {
-//                 if (item.type === 'album' && item.album) {
-//                     const alb = item.album;
-//                     return {
-//                         id: item.id,
-//                         type: 'album',
-//                         albumId: alb.album_id,
-//                         title: alb.title,
-//                         album_cover_url: alb.album_cover_url,
-//                         artist: alb.artists,
-//                         tracks: alb.songs.map((as: any) => ({
-//                             songId: as.song_id,
-//                             title: as.title,
-//                             album_cover_url: alb.album_cover_url,
-//                             preview_url: previewMap.get(as.song_id) || null,
-//                             artist: as.artists,
-//                             isrc: as.isrc
-//                         }))
-//                     };
-//                 } else {
-//                     const s = item.song;
-//                     return {
-//                         id: item.id,
-//                         type: 'song',
-//                         songId: s.song_id,
-//                         title: s.title,
-//                         album: s.album,
-//                         albumId: s.album_id,
-//                         album_cover_url: s.album?.album_cover_url,
-//                         artist: s.artists,
-//                         preview_url: previewMap.get(s.song_id) || null,
-//                         isrc: s.isrc
-//                     };
-//                 }
-//             }),
-//         };
-//     } catch (error) {
-//         console.error("Error fetching playlist:", error);
-//         return null;
-//     }
-// };
-
-export const getPlaylistById = async (playlistId: string): Promise<Playlist | null> => {
-    try {
-        const { data, error } = await supabase
-            .from('playlists')
-            .select(`
-                playlist_id,
-                name,
-                description,
-                created_at,
-                cover_url,
-                playlist_creators (
-                    user:users (user_id, name) 
-                ),
-                playlist_items (
-                    id,
-                    order,
-                    type,
-                    song:songs (
-                        album:albums(album_cover_url, title), 
-                        title, 
-                        song_id, 
-                        album_id, 
-                        isrc, 
-                        artists:artists (*)
-                    ),
-                    album:albums (
-                        album_id,
-                        title,
-                        album_cover_url,
-                        artists:artists (*),
-                        songs:songs(count)
-                    )
-                )
-            `)
-            .eq('playlist_id', playlistId)
-            .order('order', { referencedTable: 'playlist_items', ascending: true })
-            .single();
-
-        if (error || !data) throw error;
-
-        const resolvedCreators = await Promise.all(
-            data.playlist_creators.map(async (pc: any) => ({
-                ...pc.user,
-                imageUrl: pc.user.user_id ? await getProfileUrl(pc.user.user_id) : "/assets/default-avatar.png",
-                accountId: pc.user.user_id,
-                name: pc.user.name
-            }))
-        );
-
-        let songCount = 0;
-        let albumCount = 0;
-        let totalTracks = 0;
-
-        const mappedItems = data.playlist_items.map((item: any) => {
-            if (item.type === 'song') {
-                songCount++;
-                totalTracks++;
-                return {
-                    id: item.id,
-                    type: 'song' as const,
-                    songId: item.song.song_id,
-                    title: item.song.title,
-                    album: item.song.album,
-                    album_cover_url: item.song.album?.album_cover_url,
-                    artist: item.song.artists,
-                    isrc: item.song.isrc,
-                    preview_url: null 
-                };
-            } else {
-                albumCount++;
-                // Extract count from the songs(count) subquery
-                const trackCount = item.album?.songs?.[0]?.count || 0;
-                totalTracks += trackCount;
-                
-                return {
-                    id: item.id,
-                    type: 'album' as const,
-                    albumId: item.album.album_id,
-                    title: item.album.title,
-                    album_cover_url: item.album.album_cover_url,
-                    artist: item.album.artists,
-                    trackCount: trackCount,
-                    tracks: [] 
-                };
-            }
-        });
-
-        console.log("Mapped Items:", mappedItems);
-
-        return {
-            playlistId: data.playlist_id,
-            name: data.name,
-            description: data.description,
-            createdAt: data.created_at,
-            coverUrl: data.cover_url,
-            creators: resolvedCreators,
-            items: mappedItems,
-            songCount,
-            albumCount,
-            totalTracks
-        };
-    } catch (error) {
-        console.error("Error fetching playlist metadata:", error);
-        return null;
-    }
-};
-
-
-export const getAlbumTracks = async (albumId: string) => {
-    try {
-        const { data, error } = await supabase
-            .from('songs')
-            .select(`
-                song_id,
-                title,
-                isrc,
-                artists:artists (*),
-                album:albums (album_cover_url)
-            `)
-            .eq('album_id', albumId);
-
-        if (error) throw error;
-
-        // Prepare for enrichment
-        const trackInputs = data.map(s => ({
-            songId: s.song_id,
-            title: s.title,
-            artist: s.artists.map((a: any) => a.name).join(", "),
-            isrc: s.isrc
-        }));
-
-        // Only call the Edge Function for this specific album
-        const { data: enrichmentData } = await supabase.functions.invoke('enrich-album', {
-            body: { tracks: trackInputs }
-        });
-
-        const previewMap = new Map();
-        enrichmentData?.tracks?.forEach((t: any) => previewMap.set(t.songId || t.song_id, t.preview_url));
-
-        return data.map(s => ({
-            songId: s.song_id,
-            title: s.title,
-            artist: s.artists,
-            isrc: s.isrc,
-            album_cover_url: s.album[0]?.album_cover_url,
-            preview_url: previewMap.get(s.song_id) || null
-        }));
-    } catch (error) {
-        console.error("Error fetching album tracks:", error);
-        return [];
-    }
-};
-
-export async function deletePlaylist(playlistId: string) {
-    try {
-        // 1. Delete from DB and get the cover_url back
-        const { data, error } = await supabase
-            .from('playlists')
-            .delete()
-            .eq('playlist_id', playlistId)
-            .select('cover_url')
-            .single();
-
-        if (error) throw error;
-
-        const isStorageUrl = data?.cover_url?.includes('storage/v1/object/public/playlists/');
-
-        if (data?.cover_url && isStorageUrl) {
-            const parts = data.cover_url.split('/playlists/');
-            const filePath = parts[1]; 
-
-            if (filePath) {
-                const { error: storageError } = await supabase
-                    .storage
-                    .from('playlists')
-                    .remove([filePath]);
-
-                if (storageError) {
-                    console.error("Storage cleanup failed:", storageError.message);
-                }
-            }
-        }
-
-        return true;
-    } catch (err) {
-        console.error("Failed to delete playlist:", err);
-        return false;
-    }
-}
-
-
-// Generic add function for both types
-export async function addItemToPlaylist(
-    playlistId: string, 
-    itemId: string, 
-    type: 'song' | 'album', 
-    index: number
-) {
-    const insertData: any = {
-        id: itemId,
-        playlist_id: playlistId,
-        order: index,
-        type: type
-    };
-
-    if (type === 'song') insertData.song_id = itemId;
-    else insertData.album_id = itemId;
-
-    const { error } = await supabase
-        .from('playlist_items')
-        .insert(insertData);
-
-    if (error) {
-        if (error.code === '23505') return { success: true, message: 'Already in playlist' };
-        throw error;
-    }
-
-    return { success: true };
-}
-
-export async function batchAddItemsToPlaylist(items: { 
-    id: string,
-    playlist_id: string, 
-    song_id: string, 
-    type: string, 
-    order: number 
-}[]) {
-    try {
-        const { data, error } = await supabase
-            .from("playlist_items")
-            .insert(items);
-
-        if (error) throw error;
-        return { success: true, data };
-    } catch (error) {
-        console.error("Error batch adding to playlist:", error);
-        throw error;
-    }
-}
-
-// Update the remove function to handle the generic item ID
-export async function removeItemFromPlaylist(playlistId: string, itemId: string) {
-    try {
-        // Find by the row ID (playlist_items.id) rather than song_id
-        const { data: itemData, error: fetchError } = await supabase
-            .from('playlist_items')
-            .select('"order"')
-            .eq('id', itemId)
-            .single();
-
-        if (fetchError) throw fetchError;
-
-        const { error: deleteError } = await supabase
-            .from('playlist_items')
-            .delete()
-            .eq('id', itemId);
-
-        if (deleteError) throw deleteError;
-
-        await supabase.rpc('reorder_items_on_delete', {
-            p_playlist_id: playlistId,
-            p_deleted_order: itemData.order
-        });
-
-        return true;
-    } catch (err) {
-        console.error("Failed to remove item:", err);
-        return false;
-    }
-}
-
-export async function getPlaylists(userId: string): Promise<Playlist[]> {
-    try {
-        const { data: playlists, error } = await supabase
-            .from("playlists")
-            .select(`
-                playlist_id,
-                name,
-                description,
-                created_at,
-                cover_url,
-                playlist_creators!inner (user_id),
-                playlist_items (
-                    type,
-                    songs (song_id),
-                    albums (
-                        album_id,
-                        songs (song_id)
-                    )
-                )
-            `)
-            .eq("playlist_creators.user_id", userId)
-            .order("created_at", { ascending: false });
-
-        if (error) throw error;
-        if (!playlists) return [];
-
-        const results = await Promise.all(
-            playlists.map(async (p: any) => {
-                const items = p.playlist_items || [];
-                
-                const albumCount = items.filter((i: any) => i.type === 'album').length;
-                
-                const soloSongCount = items.filter((i: any) => i.type === 'song').length;
-
-                const totalTracks = items.reduce((acc: number, item: any) => {
-                    if (item.type === 'album') {
-                        return acc + (item.albums?.songs?.length || 0);
-                    }
-                    return acc + 1;
-                }, 0);
-
-                return {
-                    playlistId: p.playlist_id,
-                    name: p.name,
-                    description: p.description,
-                    createdAt: p.created_at,
-                    creators: [{
-                        accountId: userId,
-                        username: "",
-                        name: "", 
-                        imageUrl: await getProfileUrl(userId), 
-                        email: "", 
-                        bio: ""
-                    }],
-                    coverUrl: p.cover_url,
-                    items: [], // Keeping this empty as per your original structure
-                    albumCount: albumCount,
-                    songCount: soloSongCount,
-                    totalTracks: totalTracks
-                };
-            })
-        );
-
-        return results;
-    } catch (err) {
-        console.error("Failed to fetch user playlists:", err);
-        return [];                          
-    }
-}
-
-export function processPlaylistCover(file: File, size = 800): Promise<Blob> {
-    return new Promise((resolve, reject) => {
-        const img = new Image();
-        img.src = URL.createObjectURL(file);
-        
-        img.onerror = () => reject(new Error("Failed to load image"));
-
-        img.onload = () => {
-            const canvas = document.createElement("canvas");
-            canvas.width = size;
-            canvas.height = size;
-            const ctx = canvas.getContext("2d")!;
-
-            // --- CENTER CROP LOGIC ---
-            let sourceX = 0;
-            let sourceY = 0;
-            let sourceWidth = img.width;
-            let sourceHeight = img.height;
-
-            if (img.width > img.height) {
-                sourceWidth = img.height;
-                sourceX = (img.width - img.height) / 2;
-            } else {
-                sourceHeight = img.width;
-                sourceY = (img.height - img.width) / 2;
-            }
-
-            ctx.drawImage(
-                img, 
-                sourceX, sourceY, sourceWidth, sourceHeight, 
-                0, 0, size, size                             
-            );
-
-            canvas.toBlob((blob) => {
-                if (blob) resolve(blob);
-                else reject(new Error("Canvas to Blob failed"));
-            }, "image/webp", 0.85); 
-        };
-    });
-}
-
-export async function createPlaylist(
-    userId: string, 
-    name: string, 
-    description: string, 
-    imageFile: File | null
-): Promise<{ success: boolean; playlistId?: string; error?: string }> {
-    try {
-        const now = new Date();
-        const isoString = now.toISOString();
-        // 1. Generate a unique ID for the playlist first
-        const playlistId = crypto.randomUUID();
-        let coverUrl = null;
-
-        if (imageFile) {
-            const fileName = `${userId}/${playlistId}.webp`;
-            const { data: _storageData, error: storageError } = await supabase.storage
-                .from("playlists")
-                .upload(fileName, imageFile, {
-                cacheControl: '3600',
-                upsert: true
-                });
-
-            if (storageError) throw storageError;
-
-            // Get the Public URL for the uploaded image
-            const { data: urlData } = supabase.storage
-                .from("playlists")
-                .getPublicUrl(fileName);
-                
-            coverUrl = urlData.publicUrl;
-        }
-
-        // 3. Insert Playlist Row
-        const { error: playlistError } = await supabase
-        .from("playlists")
-        .insert({
-            playlist_id: playlistId,
-            name: name,
-            description: description,
-            created_at: isoString,
-            cover_url: coverUrl || "/assets/icons/music-placeholder.png"
-        });
-
-        if (playlistError) throw playlistError;
-
-        // 4. Add Creator to playlist_creators (Join Table)
-        const { error: creatorError } = await supabase
-        .from("playlist_creators")
-        .insert({
-            playlist_id: playlistId,
-            user_id: userId,
-        });
-
-        if (creatorError) throw creatorError;
-
-        return { success: true, playlistId };
-
-    } catch (error: any) {
-        console.error("Create Playlist Error:", error.message);
-        return { success: false, error: error.message };
-    }
-}
-
-export async function updatePlaylistMetadata(
-    playlistId: string, 
-    updates: { name: string; description: string }
-) {
-    const { data, error } = await supabase
-        .from('playlists')
-        .update({
-            name: updates.name,
-            description: updates.description,
-        })
-        .eq('playlist_id', playlistId)
-        .select()
-        .single();
-
-    if (error) throw error;
-    return data;
-}
-
-export async function addSongToPlaylist(playlistId: string, songId: string, index: number) {
-    const { error } = await supabase
-        .from('playlist_items')
-        .insert({
-            playlist_id: playlistId,
-            song_id: songId,
-            order: index
-        });
-
-    if (error) {
-        // Handle unique constraint if the song is already in the playlist
-        if (error.code === '23505') return { success: true, message: 'Already in playlist' };
-        throw error;
-    }
-
-    return { success: true };
-}
-
-export async function updatePlaylistCover(
-    userId: string, 
-    playlistId: string, 
-    imageFile: File
-): Promise<string> {
-    const fileName = `${userId}/${playlistId}.webp`;
-
-    // 1. Upload/Overwrite to Storage
-    const { error: storageError } = await supabase.storage
-        .from("playlists")
-        .upload(fileName, imageFile, {
-            cacheControl: '0',
-            upsert: true
-        });
-
-    if (storageError) throw storageError;
-
-    // 2. Get Public URL
-    const { data: urlData } = supabase.storage
-        .from("playlists")
-        .getPublicUrl(fileName);
-        
-    const publicUrl = `${urlData.publicUrl}?t=${Date.now()}`;
-
-    // 3. Update the Database row
-    const { error: dbError } = await supabase
-        .from("playlists")
-        .update({ cover_url: publicUrl })
-        .eq('playlist_id', playlistId);
-
-    if (dbError) throw dbError;
-
-    return publicUrl;
-}
-
-export async function updateSongsOrder(playlistId: string, orderedSongIds: {songId:any, id: any}[]) {
-    // We update each song-playlist link with its new index
-    const updates = orderedSongIds.map((songId, id, index) => ({
-        id: id,
-        playlist_id: playlistId,
-        song_id: songId,
-        order: index
-    }));
-
-    const { error } = await supabase
-        .from('playlist_items')
-        .upsert(updates, { onConflict: 'playlist_id, song_id' });
-
-    if (error) throw error;
-}
-
-export async function updateItemsOrder(playlistId: string, orderedItems: any[]) {
-    const updates = orderedItems.map((item, index) => ({
-        id: item.id, // Using the unique primary key of the join table
-        playlist_id: playlistId,
-        song_id: item.type === 'song' ? item.songId : null,
-        album_id: item.type === 'album' ? item.albumId : null,
-        type: item.type,
-        order: index
-    }));
-
-    const { error } = await supabase
-        .from('playlist_items')
-        .upsert(updates);
-
-    if (error) throw error;
-}
-
-export async function getInternalSuggestions(
-    currentAlbumId: string, 
-    userId: string | undefined, 
-    limit: number = 4
-): Promise<SuggestedAlbum[]> {
-    try {
-        let query = supabase
-        .from('album_listens')
-        .select(`
-            album_id,
-            albums!inner (
-                title,
-                album_cover_url,
-                artistalbum!inner (
-                    artists!inner (
-                        name
-                    )
-                )
-            )
-        `)
-        .neq('album_id', currentAlbumId);
-
-        if (userId) {
-            const { data: listenedIds } = await supabase
-                .from('album_listens')
-                .select('album_id')
-                .eq('user_id', userId);
-            
-            if (listenedIds && listenedIds.length > 0) {
-                const ids = listenedIds.map(l => l.album_id);
-                // Wrap IDs in quotes if they are strings/UUIDs
-                query = query.not('album_id', 'in', `(${ids.join(',')})`);
-            }
-        }
-
-        const { data, error } = await query.order('listen_date', { ascending: false }).limit(20);
-
-        if (error) throw error;
-        if (!data) return [];
-
-        const formatted: SuggestedAlbum[] = data.map(item => {
-            // 1. Check if albums is an array or object
-            const album = Array.isArray(item.albums) ? item.albums[0] : item.albums;
-            if (!album) return null;
-
-            // 2. Check if artistalbum is an array or object
-            const artistAlbumEntry = Array.isArray(album.artistalbum) 
-                ? album.artistalbum[0] 
-                : album.artistalbum;
-
-            // 3. Check if artists is an array or object (This is where your current error is)
-            const artistData = Array.isArray(artistAlbumEntry?.artists) 
-                ? artistAlbumEntry.artists[0] 
-                : artistAlbumEntry?.artists;
-
-            const artistName = artistData?.name || "Unknown Artist";
-
-            return {
-                id: item.album_id,
-                title: album.title,
-                artist_name: artistName,
-                album_cover_url: album.album_cover_url,
-            };
-        }).filter((a): a is SuggestedAlbum => a !== null);
-
-        return formatted.sort(() => Math.random() - 0.5).slice(0, limit);
-    } catch (error) {
-        console.error("Internal Suggestions Error:", error);
-        return [];
-    }
-}
