+++ src/lib/supabase/api/index.ts
@@ -1,0 +1,13 @@
+export * from "./utils/format";
+export { processPlaylistCover } from "./utils/images";
+export * from "./users";
+export * from "./social";
+export * from "./reviews";
+export * from "./catalog/songs";
+export * from "./catalog/artists";
+export * from "./catalog/albums";
+export * from "./library";
+export * from "./activities";
+export * from "./trending";
+export * from "./playlists";
+export * from "./suggestions";
