+++ tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/_auth/authlayout.tsx","./src/_auth/forgotpasswordpage.tsx","./src/_auth/resetpage.tsx","./src/_auth/forms/selectauth.tsx","./src/_auth/forms/signinform.tsx","./src/_auth/forms/signupform.tsx","./src/_root/rootlayout.tsx","./src/_root/signedinlayout.tsx","./src/_root/pages/add.tsx","./src/_root/pages/addreviewalbum.tsx","./src/_root/pages/addreviewsong.tsx","./src/_root/pages/album.tsx","./src/_root/pages/allusers.tsx","./src/_root/pages/artist.tsx","./src/_root/pages/createplaylist.tsx","./src/_root/pages/discography.tsx","./src/_root/pages/home.tsx","./src/_root/pages/library.tsx","./src/_root/pages/librarylistened.tsx","./src/_root/pages/libraryratings.tsx","./src/_root/pages/libraryreviews.tsx","./src/_root/pages/newprofile.tsx","./src/_root/pages/playlistpage.tsx","./src/_root/pages/profile.tsx","./src/_root/pages/reviewpage.tsx","./src/_root/pages/search.tsx","./src/_root/pages/settings.tsx","./src/_root/pages/songdetails.tsx","./src/_root/pages/trendingalbumspage.tsx","./src/_root/pages/trendingpage.tsx","./src/_root/pages/trendingreviewspage.tsx","./src/_root/pages/trendingsongspage.tsx","./src/_root/pages/updateprofile.tsx","./src/_root/pages/index.ts","./src/components/discographyitem.tsx","./src/components/notificationdropdown.tsx","./src/components/playlistentry.tsx","./src/components/reviewitem.tsx","./src/components/reviewitemlibrary.tsx","./src/components/suggestions.tsx","./src/components/suggestionslist.tsx","./src/components/trendinglist.tsx","./src/components/shared/authmodal.tsx","./src/components/shared/bottombar.tsx","./src/components/shared/leftsidebar.tsx","./src/components/shared/logentrymodal.tsx","./src/components/shared/playingvisualizer.tsx","./src/components/shared/playlistmodal.tsx","./src/components/shared/previewplayer.tsx","./src/components/shared/staricon.tsx","./src/components/shared/supportmodal.tsx","./src/components/shared/topbar.tsx","./src/components/shared/truncatedtext.tsx","./src/components/shared/loader.tsx","./src/components/shared/loadermusic.tsx","./src/components/ui/alert-dialog.tsx","./src/components/ui/button.tsx","./src/components/ui/dialog.tsx","./src/components/ui/form.tsx","./src/components/ui/input.tsx","./src/components/ui/label.tsx","./src/components/ui/textarea.tsx","./src/components/ui/toast.tsx","./src/components/ui/toaster.tsx","./src/constants/index.ts","./src/context/playercontext.tsx","./src/hooks/use-toast.ts","./src/lib/authcontext.tsx","./src/lib/musicboardimporter.ts","./src/lib/supabaseclient.ts","./src/lib/utils.ts","./src/lib/appwrite/api.ts","./src/lib/appwrite/deezer.ts","./src/lib/appwrite/spotify.ts","./src/lib/react-query/queryprovider.tsx","./src/lib/react-query/queriesandmutations.ts","./src/lib/react-query/querykeys.ts","./src/lib/validation/index.ts","./src/types/index.ts"],"version":"5.6.3"}
+{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/_auth/authlayout.tsx","./src/_auth/forgotpasswordpage.tsx","./src/_auth/resetpage.tsx","./src/_auth/forms/selectauth.tsx","./src/_auth/forms/signinform.tsx","./src/_auth/forms/signupform.tsx","./src/_root/rootlayout.tsx","./src/_root/signedinlayout.tsx","./src/_root/pages/add.tsx","./src/_root/pages/addreviewalbum.tsx","./src/_root/pages/addreviewsong.tsx","./src/_root/pages/album.tsx","./src/_root/pages/allusers.tsx","./src/_root/pages/artist.tsx","./src/_root/pages/createplaylist.tsx","./src/_root/pages/discography.tsx","./src/_root/pages/home.tsx","./src/_root/pages/library.tsx","./src/_root/pages/librarylistened.tsx","./src/_root/pages/libraryratings.tsx","./src/_root/pages/libraryreviews.tsx","./src/_root/pages/newprofile.tsx","./src/_root/pages/playlistpage.tsx","./src/_root/pages/profile.tsx","./src/_root/pages/reviewpage.tsx","./src/_root/pages/search.tsx","./src/_root/pages/settings.tsx","./src/_root/pages/songdetails.tsx","./src/_root/pages/trendingalbumspage.tsx","./src/_root/pages/trendingpage.tsx","./src/_root/pages/trendingreviewspage.tsx","./src/_root/pages/trendingsongspage.tsx","./src/_root/pages/updateprofile.tsx","./src/_root/pages/index.ts","./src/components/discographyitem.tsx","./src/components/notificationdropdown.tsx","./src/components/playlistentry.tsx","./src/components/reviewitem.tsx","./src/components/reviewitemlibrary.tsx","./src/components/suggestions.tsx","./src/components/suggestionslist.tsx","./src/components/trendinglist.tsx","./src/components/shared/authmodal.tsx","./src/components/shared/bottombar.tsx","./src/components/shared/leftsidebar.tsx","./src/components/shared/logentrymodal.tsx","./src/components/shared/playingvisualizer.tsx","./src/components/shared/playlistmodal.tsx","./src/components/shared/previewplayer.tsx","./src/components/shared/staricon.tsx","./src/components/shared/supportmodal.tsx","./src/components/shared/topbar.tsx","./src/components/shared/truncatedtext.tsx","./src/components/shared/loader.tsx","./src/components/shared/loadermusic.tsx","./src/components/ui/alert-dialog.tsx","./src/components/ui/button.tsx","./src/components/ui/dialog.tsx","./src/components/ui/form.tsx","./src/components/ui/input.tsx","./src/components/ui/label.tsx","./src/components/ui/textarea.tsx","./src/components/ui/toast.tsx","./src/components/ui/toaster.tsx","./src/constants/index.ts","./src/context/playercontext.tsx","./src/hooks/use-toast.ts","./src/lib/authcontext.tsx","./src/lib/supabaseclient.ts","./src/lib/utils.ts","./src/lib/integrations/deezer.ts","./src/lib/integrations/spotify.ts","./src/lib/react-query/queryprovider.tsx","./src/lib/react-query/queriesandmutations.ts","./src/lib/react-query/querykeys.ts","./src/lib/supabase/api/activities.ts","./src/lib/supabase/api/index.ts","./src/lib/supabase/api/library.ts","./src/lib/supabase/api/playlists.ts","./src/lib/supabase/api/reviews.ts","./src/lib/supabase/api/social.ts","./src/lib/supabase/api/suggestions.ts","./src/lib/supabase/api/trending.ts","./src/lib/supabase/api/users.ts","./src/lib/supabase/api/catalog/albums.ts","./src/lib/supabase/api/catalog/artists.ts","./src/lib/supabase/api/catalog/songs.ts","./src/lib/supabase/api/utils/dates.ts","./src/lib/supabase/api/utils/format.ts","./src/lib/supabase/api/utils/images.ts","./src/lib/validation/index.ts","./src/types/index.ts"],"version":"5.6.3"}
